### 2021-04-17 #3184 @kyleniemeyer
Hi @cwentland0, thanks for your interest in JOSS! This actually looks up my alley, so I'm going to serve as editor. We'll use this pre-review issue to find and assign reviewers, which I'll be doing in the coming week.
