### 2021-04-19 8660ef1
reformat docstrings for all ROM classes
