+++ perform/rom/rom_model.py
@@ -1,4 +1,5 @@
 import os
+import time
 
 import numpy as np
 
@@ -6,15 +7,42 @@ from perform.constants import REAL_TYPE
 
 
 class RomModel:
-    """
-    Base class for any ROM model
+    """Base class for all ROM models.
 
-    Assumed that every model may be equipped with
-    some sort of data standardization requirement
+    The RomModel class provides basic functionality that every ROM model should be equipped with,
+    most importantly feature scaling and mapping from the low-dimensional state to
+    the high-dimensional state ("decoding"). Feature scaling is general, but decoding operations
+    are specific to child classes. Thus, RomModel only provides the high-level utility decode_sol(),
+    which calls child class implementations of apply_decoder().
+    
+    Args:
+        model_idx: Zero-indexed ID of a given RomModel instance within a RomDomain's model_list.
+        rom_domain: RomDomain within which this RomModel is contained.
+        sol_domain: SolutionDomain with which this RomModel's RomDomain is associated.
 
-    Also assumed that every model has some means of
-    computing the full-dimensional state from the low
-    dimensional state, i.e. a "decoder"
+    Attributes:
+        model_idx: Zero-indexed ID of a given RomModel instance within a RomDomain's model_list.
+        latent_dim: Dimension of low-dimensional code associated with this model.
+        var_idxs: NumPy array of zero-indexed indices of state variables which this model maps to.
+        num_vars: Number of state variables this model maps to.
+        num_cells: Number of finite volume cells in associated SolutionDomain.
+        sol_shape: Tuple shape (num_vars, num_cells) of full-dimensional state which this model maps to.
+        target_cons: Boolean flag indicating whether this model maps to the conservative variables.
+        code: NumPy array of unsteady low-dimensional state associated with this model.
+        res: 
+            NumPy array of low-dimensional linear solve residual when integrating ROM ODE in time
+            with an implicit time integrator.
+        norm_sub_prof_cons:
+            NumPy array of conservative variable subtractive normalization profile,
+            if required by ROM method.
+        norm_fac_prof_cons:
+            NumPy array of conservative variable divisive normalization profile, if required by ROM method.
+        cent_prof_cons: NumPy array of conservative variable centering profile, if required by ROM method.
+        norm_sub_prof_prim:
+            NumPy array of primitive variable subtractive normalization profile, if required by ROM method.
+        norm_fac_prof_prim:
+        NumPy array of primitive variable divisive normalization profile, if required by ROM method.
+        cent_prof_prim: NumPy array of primitive variable centering profile, if required by ROM method.
     """
 
     def __init__(self, model_idx, rom_domain, sol_domain):
@@ -27,9 +55,8 @@ class RomModel:
         self.sol_shape = (self.num_vars, self.num_cells)
 
         # Just copy some stuff for less clutter
-        self.model_dir = rom_domain.model_dir
+        model_dir = rom_domain.model_dir
         self.target_cons = rom_domain.target_cons
-        self.hyper_reduc = rom_domain.hyper_reduc
 
         self.code = np.zeros(self.latent_dim, dtype=REAL_TYPE)
         self.res = np.zeros(self.latent_dim, dtype=REAL_TYPE)
@@ -42,21 +69,21 @@ class RomModel:
         self.cent_prof_cons = None
         self.cent_prof_prim = None
         if rom_domain.has_cons_norm:
-            self.norm_sub_prof_cons = self.load_standardization(
-                os.path.join(self.model_dir, rom_domain.norm_sub_cons_in[self.model_idx]), default="zeros"
+            self.norm_sub_prof_cons = self.load_feature_scaling(
[…]
