### 2021-04-19 0c4c5fd
docstrings for flux functions
