### 2021-04-19 3e6306b
docstrings for gas models, change gas_dict to chem_dict
