+++ perform/gas_model/gas_model.py
@@ -7,27 +7,66 @@ from perform.constants import REAL_TYPE, R_UNIV
 
 
 class GasModel:
-    """
-    Base class storing constant chemical properties of modeled species
-    Also includes universal gas methods (e.g. calc mixture molecular weight)
+    """Base class for all gas models (e.g. CPG, TPG).
+    
+    GasModel reads various universal gas properties (e.g. species molecular weights) from the chemistry input file
+    and implements several universal gas methods (e.g. calculating mixture molecular weight).
+
+    Child classes must implement the following member methods:
+
+    * calc_mix_gas_constant()
+    * calc_mix_gamma()
+    * calc_mix_cp()
+    * calc_spec_enth()
+    * calc_stag_enth()
+    * calc_species_dynamic_visc()
+    * calc_mix_dynamic_visc()
+    * calc_species_therm_cond()
+    * calc_mix_thermal_cond()
+    * calc_species_mass_diff_coeff()
+    * calc_sound_speed()
+    * calc_dens_derivs()
+    * calc_stag_enth_derivs()
+    
+    See calorically_perfect_gas.py for details on each.
+
+    Args:
+        chem_dict: Dictionary of input parameters read from chemistry file.
+
+    Attributes:
+        num_species_full: Total number of chemical species to be modeled.
+        num_species: If num_species_full == 1, num_species == 1. Otherwise, num_species == num_species_full - 1
+        species_names:
+            List of strings, names of chemical species. If not provided by user, defaults to "Species X",
+            where X is the species index.
+        mol_weights: NumPy array of chemical species molecular weights in g/mol
+        mass_frac_slice: Range for slicing mass fraction arrays, avoids issues when num_species_full == 1
+        mw_inv: NumPy array of reciprocals of molecular weights, used in some calculations
+        mw_inv_diffs:
+            NumPy array of difference in molecular weights between the first num_species species and the last species,
+            used in some calculations.
+        num_eqs:
+            Number of governing equations in system.
+            Equal to 3 + num_species (continuity, momentum, energy, and species transport).
+        mix_mass_matrix: NumPy array used in computing mixture dynamic viscosity.
+        mix_inv_mass_matrix: NumPy array used in computing mixture dynamic viscosity.
     """
 
-    def __init__(self, gas_dict):
+    def __init__(self, chem_dict):
 
         # Gas composition
-        self.num_species_full = int(gas_dict["num_species"])
-        self.mol_weights = gas_dict["mol_weights"].astype(REAL_TYPE)
+        self.num_species_full = int(chem_dict["num_species"])
+        self.mol_weights = chem_dict["mol_weights"].astype(REAL_TYPE)
 
         # Species names for plotting and output
         try:
-            self.species_names = gas_dict["species_names"]
+            self.species_names = chem_dict["species_names"]
             assert len(self.species_names) == self.num_species_full
         except KeyError:
             self.species_names = ["Species_" + str(x + 1) for x in range(self.num_species_full)]
 
         # Check input lengths
         assert len(self.mol_weights) == self.num_species_full
-        # TODO: check lengths of reaction inputs
 
         # Dealing with single-species option
         if self.num_species_full == 1:
@@ -41,34 +80,40 @@ class GasModel:
 
         self.num_eqs = self.num_species + 3
 
-        # Mass matrices for calculating
-        # 	viscosity and thermal conductivity mixing laws
+        # Mass matrices for calculating viscosity and thermal conductivity mixing laws
         self.mix_mass_matrix = np.zeros((self.num_species_full, self.num_species_full), dtype=REAL_TYPE)
         self.mix_inv_mass_matrix = np.zeros((self.num_species_full, self.num_species_full), dtype=REAL_TYPE)
-        self.precomp_mix_mass_matrices()
-
[…]
