+++ perform/limiter/limiter.py
@@ -2,8 +2,11 @@ import numpy as np
 
 
 class Limiter:
-    """
-    Base class for gradient limiters
+    """Base class for gradient limiters.
+
+    Simply provides some member functions which may be common to multiple child class limiters. 
+
+    All child classes must implement the member method calc_limiter().
     """
 
     def __init__(self):
@@ -11,8 +14,15 @@ class Limiter:
         pass
 
     def calc_neighbor_minmax(self, sol):
-        """
-        Find minimum and maximum of cell state and neighbor cell state
+        """Find minimum and maximum of cell state and neighbor cell states
+        
+        Args:
+            sol:
+                2D NumPy array of a solution profile (e.g. SolutionPhys.sol_prim),
+                where the first axis are the solution variables and the second axis are the spatial locations.
+
+        Returns:
+            NumPy arrays of the neighbor minimums and neighbor maximums, both of the same shape as sol.
         """
 
         # max and min of cell and neighbors
