+++ perform/reaction_model/finite_rate_irrev_reaction.py
@@ -6,9 +6,30 @@ from perform.reaction_model.reaction_model import ReactionModel
 
 
 class FiniteRateIrrevReaction(ReactionModel):
-    """
-    Finite rate Arrhenius reaction model assuming
-    irreversible reactions (i.e. only forwards)
+    """Finite rate Arrhenius reaction model assuming irreversible reactions.
+
+    This reaction model only computes a forward reaction rate from the Arrhenius model.
+
+    Args:
+        gas: GasModel associated with the SolutionDomain with which this ReactionModel is associated.
+        chem_dict: Dictionary of input parameters read from chemistry file.
+
+    Attributes:
+        num_reactions: Number of reactions to model.
+        nu:
+            2D NumPy array of stoichiometric coefficients, where the first axis corresponds to each reaction
+            and the second axis corresponds to each of the num_species_full chemical species.
+        nu_arr:
+            2D NumPy array of Arrhenius rate molar concentration exponential terms,
+            where the first axis corresponds to each reaction, and the second axis corresponds
+            to each of the num_species_full chemical species.
+        pre_exp_fact: 1D NumPy array of Arrhenius rate pre-exponential factors for each reaction.
+        temp_exp: 1D NumPy array of Arrhenius rate temperature exponential factors for each reaction.
+        act_energy: 1D NumPy array of Arrhenius rate activation energy for each reaction.
+        mol_weight_nu:
+            2D NumPy array of stoichiometric coefficients multiplied by molecular weight, where the first axis
+            corresponds to each reaction and the second axis corresponds to each of the num_species_full
+            chemical species. Precomputed for cost savings.
     """
 
     def __init__(self, gas, chem_dict):
@@ -38,11 +59,21 @@ class FiniteRateIrrevReaction(ReactionModel):
         self.mol_weight_nu = gas.mol_weights[None, :] * self.nu
 
     def calc_source(self, sol, dt, samp_idxs=np.s_[:]):
-        """
-        Compute chemical source term
+        """Compute reaction source term.
+
+        Args:
+            sol: SolutionPhys for which the source term is to be calculated.
+            dt:
+                Physical time step, used for thresholding rate of progress to avoid
+                consuming more mass than exists at a point.
+
+        Returns:
+            source: NumPy array of source term profiles for the num_species species transport governing equations.
+            wf: NumPy array of rate-of-progress profiles for each reaction.
         """
 
         # TODO: for larger mechanisms, definitely a lot of wasted flops where nu = 0.0
+        # TODO: wf should be stored in sol here, not returned, as this isn't general to every reaction model.
 
         gas = sol.gas_model
 
@@ -50,14 +81,14 @@ class FiniteRateIrrevReaction(ReactionModel):
         rho = sol.sol_cons[[0], samp_idxs]
         rho_mass_frac = rho * sol.mass_fracs_full[:, samp_idxs]
 
-        # rate-of-progress
+        # Rate-of-progress
         pre_exp = self.pre_exp_fact[:, None] * np.power(temp[None, :], self.temp_exp[:, None])
         wf = pre_exp * np.exp((-self.act_energy[:, None] / R_UNIV) / temp[None, :])
         wf *= np.product(
             np.power((rho_mass_frac / gas.mol_weights[:, None])[None, :, :], self.nu_arr[:, :, None]), axis=1
         )
 
-        # threshold
+        # Threshold
         wf = np.minimum(wf, rho_mass_frac[gas.mass_frac_slice, :] / dt)
 
[…]
