### 2021-04-20 ec1b636
docstrings for Solution classes and member functions
