+++ perform/solution/solution_phys.py
@@ -4,8 +4,44 @@ from perform.constants import REAL_TYPE, HUGE_NUM
 
 
 class SolutionPhys:
-    """
-    Base class for physical solution (opposed to ROM solution)
+    """Base class for full-dimensional physical solution.
+
+    This class provides the attributes and member functions required for describing a discrete system solution,
+    including transport properties, thermodynamic properties, and derivative thereof. This is general to any required
+    solution representation and is not restricted to cell-centered values. For example, this may be used to represent
+    face reconstruction states.
+
+    Args:
+        gas: GasModel associated with the SolutionDomain with which this SolutionPhys is associated.
+        num_cells: Number of finite volume cells represented by this SolutionPhys.
+        sol_prim_in: NumPy array of the primitive state profiles that this SolutionPhys represents.
+        sol_cons_in: NumPy array of the conservative state profiles that this SolutionPhys represents.
+
+    Attributes:
+        gas_model: GasModel associated with the SolutionDomain with which this SolutionPhys is associated.
+        num_cells: Number of finite volume cells represented by this SolutionPhys.
+        sol_prim: NumPy array of the primitive state profiles that this SolutionPhys represents.
+        sol_cons: NumPy array of the conservative state profiles that this SolutionPhys represents.
+        mass_fracs_full: NumPy array of all num_species_full mass fraction profiles.
+        mw_mix: NumPy array of the mixture molecular weight profile, in g/mol.
+        r_mix: NumPy array of the mixture specific gas constant profile, in J/kg-K.
+        gamma_mix: NumPy array of the mixture ratio of specific heats profile.
+        enth_ref_mix: NumPy array of the mixture reference enthalpy profile, in J/kg.
+        cp_mix: NumPy array of the mixture specific heat capacity at constant pressure profile, in J/kg-K.
+        h0: NumPy array of the stangation enthalpy profile, in J/kg.
+        hi: NumPy array of the species enthalpy profiles for all num_species_full chemical species, in J/kg.
+        c: NumPy array of the sound speed profile, in m/s.
+        dyn_visc_mix: NumPy array of the mixture dynamic viscosity profile, in N-s/m^2.
+        therm_cond_mix: NumPy array of the mixture thermal conductivity, in W/m-K.
+        mass_diff_mix:
+            NumPy array of the mass diffusivity profiles for all num_species_full chemical species, in m^2/s.
+        d_rho_d_press: NumPy array of derivative of density w/r/t pressure profile.
+        d_rho_d_temp: NumPy array of derivative of density w/r/t temperature profile.
+        d_rho_d_mass_frac: NumPy array of derivative of density w/r/t the first num_species mass fraction profiles.
+        d_enth_d_press: NumPy array of derivative of stagnation enthalpy w/r/t pressure profile.
+        d_enth_d_temp: NumPy array of derivative of stagnation enthalpy w/r/t temperature profile.
+        d_enth_d_mass_frac:
+            NumPy array of derivative of stagnation enthalpy w/r/t the first num_species mass fraction profiles.
     """
 
     def __init__(self, gas, num_cells, sol_prim_in=None, sol_cons_in=None):
@@ -14,31 +50,31 @@ class SolutionPhys:
 
         self.num_cells = num_cells
 
-        # primitive and conservative state
+        # Primitive and conservative state
         self.sol_prim = np.zeros((self.gas_model.num_eqs, num_cells), dtype=REAL_TYPE)
[…]

[…]
