### 2021-04-20 772eb48
time integrator docstrings
