### 2021-04-20 1b4ba2e
docstrings for visualizations
