+++ perform/rom/rom_domain.py
@@ -344,7 +344,7 @@ class RomDomain:
 
         sol_domain.sol_int.res_norm_l2 = norm_l2
         sol_domain.sol_int.resNormL1 = norm_l1
-        sol_domain.sol_int.res_norm_history[solver.iter - 1, :] = [norm_l2, norm_l1]
+        sol_domain.sol_int.res_norm_hist[solver.iter - 1, :] = [norm_l2, norm_l1]
 
     def set_model_flags(self):
         """Set universal ROM method flags that dictate various execution behaviors.
