+++ perform/visualization/probe_plot.py
@@ -5,11 +5,41 @@ import numpy as np
 
 from perform.visualization.visualization import Visualization
 
-# TODO: maybe easier to make probe/residual plots under some pointPlot class
+# TODO: maybe easier to make probe/residual plots under some PointPlot class
 # TODO: move some of the init input arguments used for assertions outside
 
 
 class ProbePlot(Visualization):
+    """Class for probe plot visualization.
+
+    Produces plots of the time history of probe monitors at the interval specified by vis_interval.
+
+    Must be manually expanded to plot specific profiles.
+
+    Args:
+        image_output_dir: Base image output directory, within the working directory.
+        vis_id: Zero-indexed ID of this Visualization.
+        sim_type: "FOM" or "ROM", depending on simulation type.
+        probe_vars: List of strings of variables to be probed at each probe monitor location.
+        vis_vars: List of num_subplots strings of variables to be visualized.
+        probe_num: One-indexed probe number.
+        num_probes: Total number of probe monitors set in solver_params.inp
+        vis_x_bounds:
+            List of num_subplots lists, each of length 2 including lower and upper bounds on the
+            x-axis for each subplot. Setting to [None, None] allows for dynamic axis sizing.
+        vis_y_bounds:
+            List of num_subplots lists, each of length 2 including lower and upper bounds on the
+            y-axis for each subplot. Setting to [None, None] allows for dynamic axis sizing.
+        species_names: List of strings of names for all num_species_full chemical species.
+
+    Attributes:
+        vis_type: Set to "probe".
+        x_label: Set to "t (s)", x-axis label.
+        probe_num: One-indexed probe number, corresponding to labels given in solver_params.inp.
+        probe_vars: List of strings of variables to be probed at each probe monitor location.
+        ax_line: List of matplotlib.lines.Line2D artists for each subplot.
+        fig_file: 
+    """
     def __init__(
         self,
         image_output_dir,
@@ -30,30 +60,37 @@ class ProbePlot(Visualization):
 
         self.probe_num = probe_num
         self.probe_vars = probe_vars
-        assert self.probe_num >= 0, "Must provide positive integer probe number for probe plot " + str(self.vis_id)
-        assert self.probe_num < num_probes, "probe_num_" + str(self.vis_id) + " must correspond to a valid probe"
+        assert self.probe_num > 0, "Must provide positive integer probe number for probe plot " + str(self.vis_id)
+        assert self.probe_num <= num_probes, "probe_num_" + str(self.vis_id) + " must correspond to a valid probe"
         assert vis_vars[0] is not None, "Must provide vis_vars for probe plot " + str(self.vis_id)
 
         super().__init__(vis_id, vis_vars, vis_x_bounds, vis_y_bounds, species_names)
 
         self.ax_line = [None] * self.num_subplots
 
-        # image file on disk
+        # Image file on disk
         vis_name = ""
         for vis_var in self.vis_vars:
             vis_name += "_" + vis_var
         fig_name = "probe" + vis_name + "_" + str(self.probe_num) + "_" + sim_type + ".png"
         self.fig_file = os.path.join(image_output_dir, fig_name)
 
-        # check that requested variables are being probed
+        # Check that requested variables are being probed
         for vis_var in self.vis_vars:
             assert vis_var in probe_vars, "Must probe " + vis_var + " to plot it"
 
     def plot(self, probe_vals, time_vals, iter_num, line_style, first_plot):
-        """
-        Draw and display probe plot
-        Since shape of plotted data changes every time, can't use set_data
-        As a result, probe plotting can be pretty slow
+        """Draw and display probe plot.
+        
[…]

[…]
