### 2021-04-20 9a51d34
docstrings for all top-level functions
