+++ perform/input_funcs.py
@@ -1,3 +1,5 @@
+"""Functions for handling ingestion of input files"""
+
 import os
 import re
 
@@ -7,42 +9,64 @@ from perform.constants import REAL_TYPE
 
 
 def catch_input(in_dict, in_key, default_val):
-    """
-    Assign default values if user does not provide a certain input
+    """Handle non-list dictionary entries from parameter input files.
+    
+    Casts input value as same type as default_val.
+    Assign default values if user does not provide a given input parameter.
+    Use catch_list() if attempting to retrieve lists or lists of lists.
+
+    Args:
+        in_dict: Dictionary in input parameters, within which in_key is a key.
+        in_key: Key of parameter to retrieve from in_dict.
+        default:
+            Default value to assign at output if in_key does not exist in in_dict.
+            The type of default implicitly defines the type which the parameter
+            is cast to, if it exists in in_dict.
+    Returns:
+        Input parameter retrieved from in_dict, or default if not provided.
     """
 
     # TODO: correct error handling if default type is not recognized
-    # TODO: check against lowercase'd strings
-    # 		so that inputs are not case sensitive.
-    # 		Do this for True/False too
-    # TODO: instead of trusting user for None,
-    # 		could also use NaN/Inf to indicate int/float defaults
-    # 		without passing a numerical default
-    # 		Or could just pass the actual default type lol, that'd be easier
+    # TODO: check against lowercase'd strings so that inputs are not case sensitive.
+    # Do this for True/False too
+    # TODO: instead of trusting user for None, could also use NaN/Inf to indicate int/float defaults
+    # without passing a numerical default
+    # Or could just pass the actual default type lol, that'd be easier
 
     try:
         # If None passed as default, trust user
         if default_val is None:
-            outVal = in_dict[in_key]
+            out_val = in_dict[in_key]
         else:
             default_type = type(default_val)
-            outVal = default_type(in_dict[in_key])
+            out_val = default_type(in_dict[in_key])
     except KeyError:
-        outVal = default_val
+        out_val = default_val
 
-    return outVal
+    return out_val
 
 
 def catch_list(in_dict, in_key, default, len_highest=1):
-    """
-    Input processor for reading lists or lists of lists
-    Default defines length of lists at lowest level
+    """Handle list and list of list dictionary entries from parameter input files.
+    
+    Casts list entries of input as same type as default_val.
+    Assign default values if user does not provide a given input parameter.
+
+    Args:
+        in_dict: Dictionary in input parameters, within which in_key is a key.
+        in_key: Key of parameter to retrieve from in_dict.
+        default:
+            Default list to assign at output if in_key does not exist in in_dict.
+            The type of the list entries in default implicitly defines the type which the parameter
+            is cast to, if it exists in in_dict.
+        len_highest: Expected length of topmost list.
+
+    Returns:
+        Input parameter list retrieved from in_dict, or default if not provided.
     """
 
-    # TODO: needs to throw an error if input list of lists
-    # 	is longer than len_highest
-    # TODO: could make a recursive function probably,
-    # 	just hard to define appropriate list lengths at each level
+    # TODO: needs to throw an error if input list of lists is longer than len_highest
+    # TODO: could make a recursive function probably, just hard to define appropriate list lengths at each level
 
     list_of_lists_flag = type(default[0]) == list
 
@@ -85,8 +109,13 @@ def catch_list(in_dict, in_key, default, len_highest=1):
 
 
 def parse_value(expr):
-    """
-    Parse read text value into dict value
+    """Parse text into Python expression.
+
+    Args:
+        expr: String to be converted to Python expression (e.g. a list).
+
+    Returns:
[…]
