+++ perform/time_integrator/__init__.py
@@ -7,7 +7,7 @@ def get_time_integrator(time_scheme, param_dict):
     
     Args:
         time_scheme: String name of the requested time scheme.
-        param_dict: Dictionary of parameters read from solver_params.inp.
+        param_dict: Dictionary of parameters read from the solver parameters input file.
 
     Returns:
         TimeIntegrator object of the requested type.
