### 2021-04-20 002a8d0
check linting and make corrections, configure flake8
