+++ perform/solution/solution_interior.py
@@ -12,7 +12,7 @@ class SolutionInterior(SolutionPhys):
 
     This SolutionPhys represents the interior finite volume cells of a SolutionDomain,
     i.e. all cells except the boundary ghost cells. There is only one SolutionInterior per SolutionDomain.
-    
+
     A few constructs, such as the source term, residual, residual Jacobian, etc., are only meaningful for the interior
     cells and so are represented here specifically.
 
@@ -22,7 +22,7 @@ class SolutionInterior(SolutionPhys):
     Args:
         gas: GasModel associated with the SolutionDomain with which this SolutionPhys is associated.
         sol_prim_in: NumPy array of the primitive state profiles that this SolutionPhys represents.
-        solver: SystemSolver containing global simulation parameters. 
+        solver: SystemSolver containing global simulation parameters.
         num_cells: Number of finite volume cells represented by this SolutionPhys.
         num_reactions: Number of reactions to be modeled.
         time_int: TimeIntegrator associated with the SolutionDomain with which this SolutionInterior is associated.
@@ -165,7 +165,7 @@ class SolutionInterior(SolutionPhys):
         This Jacobian is calculated when computing the approximate residual Jacobian w/r/t the conservative
         state when dual_time == False. It is also implicitly used when calculating the Roe dissipation term,
         but is not explicitly calculated there.
-        
+
         Assumes that the stagnation enthalpy, derivatives of density,
         and derivatives of stagnation enthalpy have already been computed.
 
@@ -173,7 +173,7 @@ class SolutionInterior(SolutionPhys):
         for a detailed derivation of this matrix.
 
         Args:
-            samp_idxs: 
+            samp_idxs:
                 Either a NumPy slice or NumPy array for selecting sampled cells to compute the Jacobian at.
                 Used for hyper-reduction of projection-based reduced-order models.
 
@@ -188,16 +188,14 @@ class SolutionInterior(SolutionPhys):
 
         # Initialize Jacobian
         if type(samp_idxs) is slice:
-            num_cells = sol_int.num_cells
+            num_cells = self.num_cells
         else:
             num_cells = samp_idxs.shape[0]
         gamma_matrix_inv = np.zeros((gas.num_eqs, gas.num_eqs, num_cells))
 
         # For clarity
         rho = self.sol_cons[0, samp_idxs]
-        press = self.sol_prim[0, samp_idxs]
         vel = self.sol_prim[1, samp_idxs]
-        temp = self.sol_prim[2, samp_idxs]
         mass_fracs = self.sol_prim[3:, samp_idxs]
 
         d_rho_d_press = self.d_rho_d_press[samp_idxs]
@@ -267,7 +265,7 @@ class SolutionInterior(SolutionPhys):
 
         This Jacobian is calculated when computing the residual Jacobian w/r/t the primitive
         state when dual_time == True.
-        
+
         Assumes that the stagnation enthalpy, derivatives of density,
         and derivatives of stagnation enthalpy have already been computed.
 
@@ -275,7 +273,7 @@ class SolutionInterior(SolutionPhys):
         the solver theory documentation for a detailed derivation of this matrix.
 
         Args:
-            samp_idxs: 
+            samp_idxs:
                 Either a NumPy slice or NumPy array for selecting sampled cells to compute the Jacobian at.
                 Used for hyper-reduction of projection-based reduced-order models.
 
@@ -289,16 +287,14 @@ class SolutionInterior(SolutionPhys):
 
         # Initialize Jacobian
         if type(samp_idxs) is slice:
-            num_cells = sol_int.num_cells
+            num_cells = self.num_cells
         else:
             num_cells = samp_idxs.shape[0]
[…]
