### 2021-04-20 f850c91
fix empty plot in ninth subplot spot when plotting 7 subplots
