+++ perform/visualization/field_plot.py
@@ -113,7 +113,7 @@ class FieldPlot(Visualization):
                 lin_idx = np.ravel_multi_index(([col_idx], [row_idx]), (self.num_rows, self.num_cols))[0]
                 if (lin_idx + 1) > self.num_subplots:
                     ax_var.axis("off")
-                    break
+                    continue
 
                 y_data = self.get_y_data(sol_prim, sol_cons, source, rhs, self.vis_vars[lin_idx], gas)
 
