### 2021-04-20 c7a5ec6
fix extracting and plotting probe of last species, error in probe indexing
