+++ perform/visualization/field_plot.py
@@ -79,18 +79,13 @@ class FieldPlot(Visualization):
         if not os.path.isdir(self.img_dir):
             os.mkdir(self.img_dir)
 
-    def plot(self, sol_prim, sol_cons, source, rhs, gas, x_cell, line_style, first_plot):
+    def plot(self, sol_domain, line_style, first_plot):
         """Draw and display field plot.
 
         Saves a decent amount of time by using set_ydata instead of repeatedly clearing axes and replotting.
 
         Args:
-            sol_prim: NumPy array of the primitive state profiles from SolutionInterior.
-            sol_cons: NumPy array of the conservative state profiles from SolutionInterior.
-            source: NumPy array of the reaction source term profiles for the num_species species transport equations.
-            rhs: NumPy array of the evaluation of the right-hand side function of the semi-discrete governing ODE.
-            gas: GasModel object associated with the SolutionDomain.
-            x_cell: Coordinates of Mesh cell centers.
+            sol_domain: SolutionDomain with which this Visualization is associated.
             line_style: String containing matplotlib.pyplot line style option.
             first_plot: Boolean flag indicating whether this is the first time the plot is being drawn.
         """
@@ -115,10 +110,10 @@ class FieldPlot(Visualization):
                     ax_var.axis("off")
                     continue
 
-                y_data = self.get_y_data(sol_prim, sol_cons, source, rhs, self.vis_vars[lin_idx], gas)
+                y_data = self.get_y_data(sol_domain, self.vis_vars[lin_idx])
 
                 if first_plot:
-                    x_data = x_cell
+                    x_data = sol_domain.mesh.x_cell
                     (self.ax_line[lin_idx],) = ax_var.plot(x_data, y_data, line_style)
                     ax_var.set_ylabel(self.ax_labels[lin_idx])
                     ax_var.set_xlabel(self.x_label)
@@ -136,47 +131,47 @@ class FieldPlot(Visualization):
             self.fig.tight_layout()
         self.fig.canvas.draw()
 
-    def get_y_data(self, sol_prim, sol_cons, source, rhs, var_str, gas):
+    def get_y_data(self, sol_domain, var_str):
         """Extract plotting data from flow field domain data.
 
         New inputs and var_str options must be manually added to permit new profiles to be plotted.
 
         Args:
-            sol_prim: NumPy array of the primitive state profiles from SolutionInterior.
-            sol_cons: NumPy array of the conservative state profiles from SolutionInterior.
-            source: NumPy array of the reaction source term profiles for the num_species species transport equations.
-            rhs: NumPy array of the evaluation of the right-hand side function of the semi-discrete governing ODE.
+            sol_domain: SolutionDomain with which this Visualization is associated.
             var_str: String corresponding to variable profile to be plotted.
-            gas: GasModel object associated with the SolutionDomain.
 
         Returns:
             NumPy array of the flow field profile to be visualized.
         """
 
+        sol_int = sol_domain.sol_int
+
         if var_str == "pressure":
-            y_data = sol_prim[0, :]
+            y_data = sol_int.sol_prim[0, :]
         elif var_str == "velocity":
-            y_data = sol_prim[1, :]
+            y_data = sol_int.sol_prim[1, :]
         elif var_str == "temperature":
-            y_data = sol_prim[2, :]
+            y_data = sol_int.sol_prim[2, :]
         elif var_str == "source":
-            y_data = source[0, :]
+            y_data = sol_int.source[0, :]
         elif var_str == "density":
-            y_data = sol_cons[0, :]
+            y_data = sol_int.sol_cons[0, :]
[…]
