+++ perform/visualization/probe_plot.py
@@ -142,7 +142,7 @@ class ProbePlot(Visualization):
         """
 
         var_idx = np.squeeze(np.argwhere(self.probe_vars == var_str)[0])
-        y_data = probe_vals[self.probe_num, var_idx, :iter_num]
+        y_data = probe_vals[self.probe_num - 1, var_idx, :iter_num]
 
         return y_data
 
