### 2021-04-20 983b6c8
fix plots displaying empty black window on first display interval
