### 2021-04-20 4026f07
fix gappy plotting for non-solution fields when using hyper-reduction
