### 2021-04-21 b23d166
break up calc_res_jacob() for fitting Jacobian calcs into Pressio adapter
