### 2021-04-22 3403253
redo solution updates, chem/thermo/trans calculations
- chemical composition, thermodynamic property, and transport property updates have been rolled into full state update
	- simplifies code in solver, where individual properties were calculated manually
	- less bookkeeping, but a little more expensive
- generalized pressure and temperature calculation in solution update, now this comes from calc_press_temp_from_consv() in each GasModel
	- still need to totally remove enth_ref
