+++ perform/rom/ModelAdaption/adaptROM.py
@@ -36,14 +36,18 @@ class adaptROM():
             self.adaptROMUpdateRank         =   romDomain.adaptROMUpdateRank
 
             self.FWindow                 =   []
-            self.QWindow                 =   []
             self.residualSampleIdx       =   []
 
+            self.trueSolution = []
+            self.trueCount = 50
 
         else:
             raise ValueError("Invalid selection for adaptive ROM type")
 
     def gatherStaleCons(self, romDomain, solDomain, solver, model):
+        '''Gathers the conservative state from the stale file'''
+
+        #solDomain.solHistCons is analogous to Q
 
         assert(romDomain.staleSnapshots.shape[1] == solver.mesh.numCells), 'resolution mis-match between stale states and current simulation'
 
@@ -51,47 +55,35 @@ class adaptROM():
         solDomain.timeIntegrator.staleStatetimeOrder =  romDomain.timeIntegrator.staleStatetimeOrder
 
         self.FWindow = np.zeros((model.numVars, solver.mesh.numCells, self.adaptROMWindowSize))
-        self.QWindow = np.zeros((model.numVars, solver.mesh.numCells, romDomain.timeIntegrator.timeOrder+1))
 
         if self.adaptROMWindowSize != 1 :
-            self.FWindow[:, :, :self.adaptROMWindowSize-1] = romDomain.staleSnapshots[model.varIdxs, :, -self.adaptROMWindowSize+1 :]
+            self.FWindow[:, :, :-1] = romDomain.staleSnapshots[model.varIdxs, :, -(self.adaptROMWindowSize-1):]
 
-        self.QWindow[:, :, 1:romDomain.timeIntegrator.staleStatetimeOrder+1] = np.flip(romDomain.staleSnapshots[model.varIdxs, :, -romDomain.timeIntegrator.staleStatetimeOrder:], axis = 2)
-        self.QWindow[:, :, 0] = self.QWindow[:, :, 1].copy()
+        for timeIdx in range(1, romDomain.timeIntegrator.staleStatetimeOrder+1):
+            solDomain.solInt.solHistCons[timeIdx] = romDomain.staleSnapshots[:, :, -timeIdx].copy()
+
+        solDomain.solInt.solHistCons[0] = solDomain.solInt.solHistCons[1].copy()
+
+        self.trueSolution = np.load('/Users/sahilbhola/Documents/CASLAB/perform/examples/testCase/UnsteadyFieldResults/True_cons.npy')
 
 
     def HistoryInitialization(self, romDomain, solDomain, solver, model):
-    #     #TODO: currently solHistPrim is not being updated. Required if solving based on primitive variables...
+        '''Computes the coded state and the reconstructed state for initializing the sub-iteration'''
 
         timeOrder = min(solver.iter, romDomain.timeIntegrator.timeOrder)  # cold start
-        timeOrder = max(romDomain.timeIntegrator.staleStatetimeOrder, timeOrder) #time order update if stale states are available
+        timeOrder = max(romDomain.timeIntegrator.staleStatetimeOrder, timeOrder)  # updating time order if stale states are available
 
+        for timeIdx in range(timeOrder+1):
+            #Updating coded history
+            solCons = solDomain.solInt.solHistCons[timeIdx].copy()
 
-        for timeIdx in range(1, timeOrder + 1):
-            # reading available solution
-            solCons = self.QWindow[:, :, timeIdx].copy()
+            solCons = model.standardizeData(solCons[model.varIdxs, :], normalize=True, normFacProf=model.normFacProfCons, normSubProf=model.normSubProfCons,
+									   center=True, centProf=model.centProfCons, inverse=False)
 
-            # standardizing solution
[…]
