### 2021-06-16 1c4963a
cleaned adaptive rom
