+++ perform/rom/ModelAdaption/adaptROM.py
@@ -1,8 +1,6 @@
 import numpy as np
 from scipy.linalg import orth
-import copy
 import pdb
-import matplotlib.pyplot as plt
 
 
 class adaptROM():
@@ -16,7 +14,6 @@ class adaptROM():
             Method developed by Prof. Karthik Duraisamy (University of Michigan)
             """
             #TODO: Implement residual sampling step
-
             self.adaptsubIteration              =   True
             self.trueStandardizedState          =   np.zeros((model.numVars, solver.mesh.numCells))
             self.adaptionResidual               =   np.zeros((model.numVars*solver.mesh.numCells))
@@ -30,60 +27,71 @@ class adaptROM():
             Method developed by Prof. Benjamin Peherstorfer (NYU)
             """
 
-            self.adaptROMResidualSampStep   =   romDomain.adaptROMResidualSampStep
-            self.adaptROMnumResSamp         =   romDomain.adaptROMnumResSamp
-            self.adaptROMWindowSize         =   romDomain.adaptROMWindowSize
-            self.adaptROMUpdateRank         =   romDomain.adaptROMUpdateRank
+            self.adaptROMResidualSampStep   =   romDomain.adaptROMResidualSampStep  #specifies the number of step after which full residual is computed
+            self.adaptROMnumResSamp         =   romDomain.adaptROMnumResSamp        #specifies the number of samples of the residual
+            self.adaptROMWindowSize         =   romDomain.adaptROMWindowSize        #look-back widow size
+            self.adaptROMUpdateRank         =   romDomain.adaptROMUpdateRank        #basis upadate rank
+            self.adaptROMInitialSnap        =   romDomain.adaptROMInitialSnap       #specifies the "number" of samples available (same as the number of samples used for basis computation)
 
-            self.FWindow                 =   []
-            self.residualSampleIdx       =   []
+            self.FWindow                    =   np.zeros((model.numVars, solver.mesh.numCells, self.adaptROMWindowSize)) #look-back window
+            self.residualSampleIdx          =   [] #residual sample indexes (indexes fron all the states will be pooled together)
 
-            self.trueSolution = []
-            self.trueCount = 50
+            self.interPolAdaptionWindow     =   []
+            self.interPolBasis              =   []
+            self.scaledAdaptionWindow       =   np.zeros((model.numVars*solver.mesh.numCells, self.adaptROMWindowSize))
+
+
+            assert(self.adaptROMWindowSize-1<=self.adaptROMInitialSnap), 'Look back window size minus 1 should be less than equal to the available stale states'
 
         else:
             raise ValueError("Invalid selection for adaptive ROM type")
 
-    def gatherStaleCons(self, romDomain, solDomain, solver, model):
-        '''Gathers the conservative state from the stale file'''
+    def initializeLookBackWindow(self, romDomain, model):
 
-        #solDomain.solHistCons is analogous to Q
+        self.FWindow[:, :, :-1] = romDomain.staleConsSnapshots[model.varIdxs, :, -(self.adaptROMWindowSize-1):]
 
-        assert(romDomain.staleSnapshots.shape[1] == solver.mesh.numCells), 'resolution mis-match between stale states and current simulation'
-
-        romDomain.timeIntegrator.staleStatetimeOrder = min(romDomain.timeIntegrator.timeOrder, romDomain.staleSnapshots.shape[-1])
[…]
