+++ perform/rom/romDomain.py
@@ -92,13 +92,13 @@ class romDomain:
 		self.adaptiveROM = catchInput(romDict, "adaptiveROM", False)
 		if self.adaptiveROM:
 				self.adaptiveROMMethod 				=	catchInput(romDict, "adaptiveROMMethod", "OSAB")
-				self.adaptROMstaleConsFName	  		=	catchInput(romDict, "adaptROMstaleConsFName", "solCons_FOM.npy")
 				self.adaptROMResidualSampStep		=	catchInput(romDict, "adaptROMResidualSampStep", 1)
 				self.adaptROMnumResSamp				=	catchInput(romDict, "adaptROMnumResSamp", solver.mesh.numCells)
-				self.adaptROMUpdateRank				=	catchInput(romDict, "adaptROMUpdateRank", 1)
-				assert(self.adaptROMUpdateRank > 0), '~adaptROMUpdateRank~ should be greater than zero'
+				self.adaptROMInitialSnap			=	catchInput(romDict, "adaptROMInitialSnap", 1)
 				self.adaptROMWindowSize				=	catchInput(romDict, "adaptROMWindowSize", 1)
-				assert(self.adaptROMWindowSize > 0), '~adaptROMWindowSize~ should be greater than zero'
+				self.adaptROMUpdateRank				=	catchInput(romDict, "adaptROMUpdateRank", 1)
+				self.adaptROMConsSnapFName			=	catchInput(romDict, "adaptROMConsSnapFName", "solCons_FOM.npy")
+
 
 		# set up hyper-reduction, if necessary
 		self.hyperReduc = catchInput(romDict, "hyperReduc", False)
@@ -156,18 +156,23 @@ class romDomain:
 
 		# gather solution history from stale solution data
 		if self.adaptiveROM and self.adaptiveROMMethod=='AADEIM':
+			assert (os.path.exists(os.path.join(const.unsteadyOutputDir, self.adaptROMConsSnapFName))), \
+				'Path for stale conservative state : '+ str(os.path.join(const.unsteadyOutputDir, self.adaptROMConsSnapFName))+ ' does not exist'
 
-			assert (os.path.exists(os.path.join(const.unsteadyOutputDir, self.adaptROMstaleConsFName))), \
-				"Path : " + str(os.path.join(const.unsteadyOutputDir, self.adaptROMstaleConsFName)) + " does not exist"
+			self.staleConsSnapshots = np.load(os.path.join(const.unsteadyOutputDir, self.adaptROMConsSnapFName))
+			assert (self.staleConsSnapshots.shape[1] == solver.mesh.numCells), 'resolution mis-match between stale states and current simulation'
 
-			self.staleSnapshots = np.load(os.path.join(const.unsteadyOutputDir, self.adaptROMstaleConsFName))
+			self.timeIntegrator.staleStatetimeOrder = min(self.timeIntegrator.timeOrder,self.staleConsSnapshots.shape[-1])
+			solDomain.timeIntegrator.staleStatetimeOrder = self.timeIntegrator.staleStatetimeOrder
 
-			assert(self.staleSnapshots.shape[-1]>= self.adaptROMWindowSize), 'Insufficient stale snapshots for executing AADEIM'
+			for timeIdx in range(1, self.timeIntegrator.staleStatetimeOrder + 1):
+				solDomain.solInt.solHistCons[timeIdx] = self.staleConsSnapshots[:, :, -timeIdx].copy()
 
-			for modelIdx, model in enumerate(self.modelList): model.adapt.gatherStaleCons(self, solDomain, solver, model)
+			solDomain.solInt.solHistCons[0] = solDomain.solInt.solHistCons[1].copy()
[…]
