### 2021-06-16 4587490
updated reduced residual computation (2)
