### 2021-06-20 #3184 @cwentland0
Hi @arfon, here are a few names:
clemaitre58
szkafander
tupi
Himscipy
highlando
bonh

These reviewers indicated Python as a preferred programming language, and made some reference to computational fluid dynamics and machine learning/data-driven methods in their preferred domains/topics. If figure that combo might fit best.
