+++ perform/limiter/venkat_limiter.py
@@ -15,7 +15,7 @@ class VenkatLimiter(Limiter):
 
         super().__init__()
 
-    def calc_limiter(self, sol_domain, grad):
+    def calc_limiter(self, sol_domain, sol_full, grad):
         """Compute multiplicative limiter.
 
         Args:
@@ -26,46 +26,35 @@ class VenkatLimiter(Limiter):
             NumPy array of the multiplicative gradient limiter profile.
         """
 
-        sol_prim = sol_domain.sol_prim_full[:, sol_domain.grad_idxs]
+        sol = sol_full[:, sol_domain.grad_idxs]
 
         # get min/max of cell and neighbors
-        sol_prim_min, sol_prim_max = self.calc_neighbor_minmax(sol_domain.sol_prim_full[:, sol_domain.grad_neigh_idxs])
+        sol_min, sol_max = self.calc_neighbor_minmax(sol_full[:, sol_domain.grad_neigh_idxs])
 
         # extract gradient cells
-        sol_prim_min = sol_prim_min[:, sol_domain.grad_neigh_extract]
-        sol_prim_max = sol_prim_max[:, sol_domain.grad_neigh_extract]
+        sol_min = sol_min[:, sol_domain.grad_neigh_extract]
+        sol_max = sol_max[:, sol_domain.grad_neigh_extract]
 
         # unconstrained reconstruction at neighboring cell centers
-        d_sol_prim = grad * sol_domain.mesh.dx
-        sol_prim_left = sol_prim - d_sol_prim
-        sol_prim_right = sol_prim + d_sol_prim
+        d_sol = grad * sol_domain.mesh.dx
+        sol_left = sol - d_sol
+        sol_right = sol + d_sol
 
         # limiter defaults to 1
-        phi_left = np.ones(sol_prim.shape, dtype=REAL_TYPE)
-        phi_right = np.ones(sol_prim.shape, dtype=REAL_TYPE)
+        phi_left = np.ones(sol.shape, dtype=REAL_TYPE)
+        phi_right = np.ones(sol.shape, dtype=REAL_TYPE)
 
         # find idxs where difference is either positive or negative
-        cond1_left = (sol_prim_left - sol_prim) > 0
-        cond1_right = (sol_prim_right - sol_prim) > 0
-        cond2_left = (sol_prim_left - sol_prim) < 0
-        cond2_right = (sol_prim_right - sol_prim) < 0
+        cond1_left = (sol_left - sol) > 0
+        cond1_right = (sol_right - sol) > 0
+        cond2_left = (sol_left - sol) < 0
+        cond2_right = (sol_right - sol) < 0
 
         # apply smooth Venkatakrishnan function
-        phi_left[cond1_left] = self.venkat_function(
-            sol_prim_max[cond1_left], sol_prim[cond1_left], sol_prim_left[cond1_left]
-        )
-
-        phi_right[cond1_right] = self.venkat_function(
-            sol_prim_max[cond1_right], sol_prim[cond1_right], sol_prim_right[cond1_right]
-        )
-
-        phi_left[cond2_left] = self.venkat_function(
-            sol_prim_min[cond2_left], sol_prim[cond2_left], sol_prim_left[cond2_left]
-        )
-
-        phi_right[cond2_right] = self.venkat_function(
-            sol_prim_min[cond2_right], sol_prim[cond2_right], sol_prim_right[cond2_right]
-        )
+        phi_left[cond1_left] = self.venkat_function(sol_max[cond1_left], sol[cond1_left], sol_left[cond1_left])
+        phi_right[cond1_right] = self.venkat_function(sol_max[cond1_right], sol[cond1_right], sol_right[cond1_right])
[…]
