### 2021-06-22 #3184 @kyleniemeyer
Hi @clemaitre58 and @tupi, would you be willing to review this submission for JOSS?
