### 2021-06-23 f5bcf81
basic structure of non-intrusive ROMs, totally broken right now
