+++ perform/rom/nonintrusive_rom/autoencoder_rom/autoencoder_koopman/pan2020.py
@@ -1,0 +1,8 @@
+from perform.rom.nonintrusive_rom.autoencoder_rom.autoencoder_koopman.autoencoder_koopman import AutoencoderKoopman
+
+
+class AEKoopmanPan2020(AutoencoderKoopman):
+    """Class implementing the continuous-time variant of the autoencoder Koopman method by Pan (2020).
+
+
+    """
