+++ perform/rom/nonintrusive_rom/autoencoder_rom/autoencoder_koopman/otto2019.py
@@ -1,0 +1,8 @@
+from perform.rom.nonintrusive_rom.autoencoder_rom.autoencoder_koopman.autoencoder_koopman import AutoencoderKoopman
+
+
+class AEKoopmanOtto2019(AutoencoderKoopman):
+    """Class implementing the discrete-time variant of the autoencoder Koopman method by Otto (2019).
+
+    
+    """
