+++ perform/rom/ml_library/ml_library.py
@@ -1,0 +1,11 @@
+
+
+class MLLibrary():
+    """Base class for abstracting machine learning library functionality.
+
+
+    """
+
+    def __init__():
+
+        pass
