### 2021-06-23 #3184 @clemaitre58
ok

On Tue, Jun 22, 2021 at 10:05 PM Kyle Niemeyer ***@***.***>
wrote:

> Hi @clemaitre58 <https://github.com/clemaitre58> and @tupi
> <https://github.com/tupi>, would you be willing to review this submission
> for JOSS?
>
