### 2021-06-23 #3184 @kyleniemeyer
Thanks @clemaitre58! Once we find a second reviewer, I will initiate the actual review issue. In the meantime, you can see our reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
