### 2021-06-24 e755636
Big reorganizing to make ROM models more modular w/r/t ML libraries
- No more subclasses for ROM models for a given ML library (e.g. autoencoder_galerkin_proj_tfkeras)
- Now, MLLibrary class provides abstraction layer for accessing ML library functionality
    - child classes of MLLibrary (e.g. TFKerasLibrary) provide library-specific implementations
- Should really smooth future implementation of PyTorch
