+++ perform/rom/projection_rom/autoencoder_proj_rom/autoencoder_splsvt_proj.py
@@ -1,14 +1,14 @@
 import numpy as np
 
-from perform.rom.projection_rom.autoencoder_proj_rom.autoencoder_tfkeras.autoencoder_tfkeras import AutoencoderTFKeras
+from perform.rom.projection_rom.autoencoder_proj_rom.autoencoder_proj_rom import AutoencoderProjROM
 
 
-class AutoencoderLSPGProjTFKeras(AutoencoderTFKeras):
-    """Class for projection-based ROM with a TF-Keras non-linear manifold decoder and LSPG projection.
+class AutoencoderSPLSVTProj(AutoencoderProjROM):
+    """Class for projection-based ROM with a TF-Keras non-linear manifold decoder and SP-LSVT projection.
 
-    Inherits from AutoencoderTFKeras.
+    Inherits from AutoencoderProjROM.
 
-    Decoder is assumed to map to the conservative variables. Allows implicit time integration only.
+    Decoder is assumed to map to the primitive variables. Allows implicit time integration only.
 
     Args:
         model_idx: Zero-indexed ID of a given RomModel instance within a RomDomain's model_list.
@@ -19,32 +19,31 @@ class AutoencoderLSPGProjTFKeras(AutoencoderTFKeras):
     def __init__(self, model_idx, rom_domain, sol_domain):
 
         if rom_domain.time_integrator.time_type == "explicit":
-            raise ValueError(
-                "NLM LSPG with an explicit time integrator deteriorates to Galerkin, please use Galerkin"
-                + " or select an implicit time integrator."
-            )
+            raise ValueError("Explicit NLM SP-LSVT not implemented yet")
 
-        if rom_domain.time_integrator.dual_time:
-            raise ValueError("LSPG is intended for conservative variable evolution, please set dual_time = False")
+        if (rom_domain.time_integrator.time_type == "implicit") and (not rom_domain.time_integrator.dual_time):
+            raise ValueError(
+                "NLM SP-LSVT is intended for primitive variable  evolution, please use Galerkin or LSPG,"
+                + " or set dual_time = True"
+            )
 
         super().__init__(model_idx, rom_domain, sol_domain)
 
         if self.encoder_jacob:
             raise ValueError(
-                "LSPG is not equipped with an encoder Jacobian approximation, please set encoder_jacob = False"
+                "SP-LSVT is not equipped with an encoder Jacobian approximation, please set encoder_jacob = False"
             )
 
     def calc_d_code(self, res_jacob, res, sol_domain):
         """Compute change in low-dimensional state for implicit scheme Newton iteration.
 
         This function computes the iterative change in the low-dimensional state for a given Newton iteration
-        of an implicit time integration scheme. Please refer to Lee and Carlberg (2020) for details on the
-        formulation for LSPG projection.
+        of an implicit time integration scheme.
 
         Args:
             res_jacob:
                 scipy.sparse.csr_matrix containing full-dimensional residual Jacobian with respect to
-                the conservative variables.
+                the primitive variables.
             res: NumPy array of fully-discrete residual, already negated for Newton iteration.
             sol_domain: SolutionDomain with which this RomModel's RomDomain is associated.
 
@@ -57,8 +56,8 @@ class AutoencoderLSPGProjTFKeras(AutoencoderTFKeras):
         """
 
         # decoder Jacobian, scaled
-        jacob = self.calc_model_jacobian(sol_domain)
-        scaled_jacob = jacob * self.norm_fac_prof_cons.ravel(order="C")[:, None]
[…]
