### 2021-06-24 22d372c
clean up new ML library interface, docstrings
