+++ perform/rom/projection_rom/autoencoder_proj_rom/autoencoder_splsvt_proj.py
@@ -56,7 +56,7 @@ class AutoencoderSPLSVTProj(AutoencoderProjROM):
         """
 
         # decoder Jacobian, scaled
-        jacob = self.calc_jacobian(sol_domain, encoder_jacob=self.encoder_jacob)
+        jacob = self.calc_jacobian(sol_domain)
         scaled_jacob = jacob * self.norm_fac_prof_prim.ravel(order="C")[:, None]
 
         # test basis
