+++ perform/rom/ml_library/tfkeras_library.py
@@ -1,4 +1,5 @@
 import os
+from packaging import version
 
 import numpy as np
 import tensorflow as tf
@@ -9,16 +10,24 @@ from perform.constants import FD_STEP_DEFAULT
 
 
 class TFKerasLibrary(MLLibrary):
-    """Class for accessing Tensorflow-Keras functionalities."""
+    """Class for implementing Tensorflow-Keras functionalities.
+    
+    This class assumes that Tensorflow >=2.0 is installed. 
+
+    Args:
+        rom_domain: RomDomain which contains relevant ROM input dictionary.
+    """
 
     def __init__(self, rom_domain):
 
         super().__init__(rom_domain)
+        if version.parse(tf.__version__) < version.parse("2.0"):
+            raise ValueError("You must use TensorFlow >=2.0, please upgrade your installation of Tensorflow.")
 
     def init_device(self, run_gpu):
-        """Initializes GPU execution, if requested
+        """Initializes GPU execution, if requested.
 
-        TensorFlow 2+ can execute from CPU or GPU, this function does some prep work.
+        TensorFlow >=2.0 can execute from CPU or GPU, this function does some prep work.
 
         Passing run_gpu=True will limit GPU memory growth, as unlimited TensorFlow memory allocation can be
         quite aggressive on first call.
@@ -50,61 +59,70 @@ class TFKerasLibrary(MLLibrary):
         """Load model object from file.
 
         This function loads a trained Tensorflow-Keras model.
-        Either the newer SavedModel
+        Either the newer SavedModel format or old Keras HDF5 format are supported.
 
         Args:
             model_path: string path to the model file to be loaded.
+            custom_objects: dictionary of custom classes or functions for use in model.
 
         Returns:
-            Uncompiled Keras model object. As no training is being done, there is no need to compile.
+            Uncompiled tf.keras.Model object. As no training is being done, there is no need to compile.
         """
 
         model_obj = load_model(model_path, compile=False, custom_objects=custom_objects)
 
         return model_obj
 
-    def init_persistent_mem(self, input_shape, input_dtype="float64", prepend_batch=False):
+    def init_persistent_mem(self, data_shape, dtype="float64", prepend_batch=False):
         """Initialize persistent memory for graph computations.
 
         In order to use tf.function for efficient graph computations, some persistent memory must be allocated.
 
         Args:
-            model: child class of RomModel which provides necessary Boolean flags.
-            model_type: string denoting the appropriate types of memory to allocate for a given model.
-            sol_domain: SolutionDomain containing physical solutions.
-            code: low-dimensional latent variables.
+            data_shape: tuple shape of memory object to be created.
+            dtype: data type of memory object to be created.
+            prepend_batch: Boolean flag indicating whether to prepend a singleton batch dimension to memory object.
 
         Returns:
-            Tuple of presistent memory variables, specific to the requested model_type.
+            tf.Variable object to be used for persistent memory.
         """
 
-        mem_input = np.zeros(input_shape, dtype=input_dtype)
+        mem_np = np.zeros(data_shape, dtype=dtype)
         if prepend_batch:
-            mem_input = np.expand_dims(mem_input, axis=0)
-        mem_obj = tf.Variable(mem_input, dtype=input_dtype)
+            mem_np = np.expand_dims(mem_np, axis=0)
+        mem_obj = tf.Variable(mem_np, dtype=dtype)
 
         return mem_obj
 
     def check_conv_io_format(self, io_format):
-        """Checks for convolutional model I/O compatability."""
+        """Checks for convolutional model I/O compatability.
+        
+        Tensorflow >=2.0 cannot handle NCHW convolutional layers when running on a CPU.
[…]
