+++ perform/rom/projection_rom/autoencoder_proj_rom/autoencoder_proj_rom.py
@@ -14,38 +14,39 @@ class AutoencoderProjROM(ProjectionROM):
 
     sol = cent_prof + norm_sub_prof + norm_fac_prof * decoder(code)
 
-    Child classes must implement the following member functions with
-    library-specific (e.g. TensorFlow-Keras, PyTorch) implementations:
-    * load_model_obj()
-    * check_model()
-    * apply_decoder()
-    * apply_encoder()
-    * calc_analytical_model_jacobian()
-    * calc_numerical_model_jacobian()
-    * calc_model_jacobian()
-    Check autoencoder_tfkeras.py for further details on these methods.
-
     Args:
         model_idx: Zero-indexed ID of a given RomModel instance within a RomDomain's model_list.
         rom_domain: RomDomain within which this RomModel is contained.
         sol_domain: SolutionDomain with which this RomModel's RomDomain is associated.
 
     Attributes:
+        mllib: MLLibrary instance associated with rom_domain.
         decoder:
             Decoder model object (e.g. tf.keras.Model) which maps from the low-dimensional state
             to the full-dimensional state.
+        decoder_isconv: Boolean flag indicating whether decoder output layer is convolutional.
+        decoder_io_format:
+            Either "channels_first" (i.e. NCHW) or "channels_last" (i.e. NHWC) format of decoder output
+            convolutional layer, if decoder_isconv = True. 
         decoder_io_dtypes: List containing the data type of the decoder input and output.
+        decoder_io_shapes: tuple containing the tuple shapes of the decoder input and output.
         encoder_jacob:
             Boolean flag indicating whether the method should use an encoder Jacobian approximation, if applicable.
         encoder:
             Encoder model object (e.g. tf.keras.Model) which maps from the full-dimensional state
             to the low-dimensional state. Only required if initializing the simulation from a full-dimensional
             initial conditions or the ROM method requires an encoder.
+        encoder_isconv: Boolean flag indicating whether encoder input layer is convolutional.
+        encoder_io_format:
+            Either "channels_first" (i.e. NCHW) or "channels_last" (i.e. NHWC) format of encoder input
+            convolutional layer, if decoder_isconv = True. 
         encoder_io_dtypes: List containing the data type of the encoder input and output.
+        encoder_io_shapes: tuple containing the tuple shapes of the encoder input and output.
         numerical_jacob:
             Boolean flag indicating whether the decoder/encoder Jacobian should be computed numerically.
             If False (default), computes the Jacobian analytically using automatic differentiation.
         fd_step: Finite-difference step size for computing the numerical decoder/encoder Jacobian, if requested.
+        jacob_input: Persistent memory variable which will contain model inputs about which Jacobians are calculated.
     """
 
     def __init__(self, model_idx, rom_domain, sol_domain):
@@ -65,16 +66,18 @@ class AutoencoderProjROM(ProjectionROM):
         self.decoder_io_format = catch_input(rom_dict, "model_io_format", None)
         decoder_input_shape = (self.latent_dim,)
         if self.decoder_isconv:
-            if self.decoder_io_format == "nhwc":
+            if self.decoder_io_format == "channels_last":
                 decoder_output_shape = (
                     self.num_cells,
                     self.num_vars,
                 )
-            elif self.decoder_io_format == "nchw":
+            elif self.decoder_io_format == "channels_first":
                 decoder_output_shape = (
                     self.num_vars,
                     self.num_cells,
                 )
+            else:
[…]
