### 2021-06-28 #3184 @kyleniemeyer
Hello @Himscipy, would you be willing to review this submission for JOSS? If you haven't reviewed for us before, we follow a checklist-based approach in the open here on GitHub: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
