### 2021-06-28 #3428 @whedon
[REVIEW]: PERFORM: A Python package for developing reduced-order models for reacting fluid flows
**Submitting author:** @cwentland0 (<a href="http://orcid.org/0000-0002-8500-569X">Christopher Wentland</a>)
**Repository:** https://github.com/cwentland0/perform
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.1
