### 2021-06-29 020819c
add Barth-Jespersen face limiter, heat release calcs and output
- deviating a bit in this branch, my b
- made BarthJespLimiter an umbrella class
    - now have BarthJespCellLimiter and BarthJespFaceLimiter child classes
    - face limiter produces smooth pressure field for higher-resolution
- add heat release calculations
    - rate of progress limiting and source term calcs are now non-vectorized, need to improve
    - can now request heat release unsteady field output with hr_out
    - can now request heat release plotting with "heat_release"
