### 2021-07-15 64a4681
fix non-hyper-reduced models for separated bases
