+++ perform/rom/projection_rom/autoencoder_proj_rom/autoencoder_tfkeras/autoencoder_splsvt_proj_tfkeras.py
@@ -34,7 +34,7 @@ class AutoencoderSPLSVTProjTFKeras(AutoencoderTFKeras):
                 "SP-LSVT is not equipped with an encoder Jacobian approximation, please set encoder_jacob = False"
             )
 
-    def calc_d_code(self, res_jacob, res, sol_domain):
+    def calc_d_code(self, res_jacob, res, sol_domain, rom_domain):
         """Compute change in low-dimensional state for implicit scheme Newton iteration.
 
         This function computes the iterative change in the low-dimensional state for a given Newton iteration
@@ -48,9 +48,6 @@ class AutoencoderSPLSVTProjTFKeras(AutoencoderTFKeras):
             sol_domain: SolutionDomain with which this RomModel's RomDomain is associated.
 
         Returns:
-            d_code:
-                Solution of low-dimensional linear solve, representing the iterative change in
-                the low-dimensional state.
             lhs: Left-hand side of low-dimensional linear solve.
             rhs: Right-hand side of low-dimensional linear solve.
         """
@@ -66,6 +63,4 @@ class AutoencoderSPLSVTProjTFKeras(AutoencoderTFKeras):
         lhs = test_basis.T @ test_basis
         rhs = test_basis.T @ (res / self.norm_fac_prof_cons).ravel(order="C")
 
-        d_code = np.linalg.solve(lhs, rhs)
-
-        return d_code, lhs, rhs
+        return lhs, rhs
