+++ perform/rom/romDomain.py
@@ -92,14 +92,12 @@ class romDomain:
 		self.adaptiveROM = catchInput(romDict, "adaptiveROM", False)
 		if self.adaptiveROM:
 				self.adaptiveROMMethod 				=	catchInput(romDict, "adaptiveROMMethod", "OSAB")
-				self.adaptROMResidualSampStep		=	catchInput(romDict, "adaptROMResidualSampStep", 1)
-				self.adaptROMnumResSamp				=	catchInput(romDict, "adaptROMnumResSamp", solver.mesh.numCells)
-				self.adaptROMInitialSnap			=	catchInput(romDict, "adaptROMInitialSnap", 1)
-				self.adaptROMWindowSize				=	catchInput(romDict, "adaptROMWindowSize", 1)
-				self.adaptROMUpdateRank				=	catchInput(romDict, "adaptROMUpdateRank", 1)
-				self.adaptROMConsSnapFName			=	catchInput(romDict, "adaptROMConsSnapFName", "solCons_FOM.npy")
-
-
+				self.adaptROMResidualSampStep		=	catchInput(romDict, "adaptROMResidualSampStep", 1) #Steps after which the residual samples will get updated
+				self.adaptROMnumResSamp				=	catchInput(romDict, "adaptROMnumResSamp", solver.mesh.numCells) #Number of points at which the residual is sampled 
+				self.adaptROMInitialSnap			=	catchInput(romDict, "adaptROMInitialSnap", 1) #Numder of initial snapshots (of Cons. state) available (also used for the initial basis)
+				self.adaptROMWindowSize				=	catchInput(romDict, "adaptROMWindowSize", 1) #Look back window size
+				self.adaptROMUpdateRank				=	catchInput(romDict, "adaptROMUpdateRank", 1) #Rank of updated to the basis
+				self.adaptROMConsSnapFName			=	catchInput(romDict, "adaptROMConsSnapFName", "solCons_FOM.npy") #File name of the cons. snaps file
 		# set up hyper-reduction, if necessary
 		self.hyperReduc = catchInput(romDict, "hyperReduc", False)
 		if (self.isIntrusive and self.hyperReduc):
@@ -136,6 +134,7 @@ class romDomain:
 
 		solDomain.solInt.updateState(fromCons=self.targetCons)
 
+
 		# get time integrator, if necessary
 		# TODO: timeScheme should be specific to the romDomain, not the solver
 		if self.hasTimeIntegrator:
@@ -153,9 +152,9 @@ class romDomain:
 		solDomain.solInt.solHistCons = [solDomain.solInt.solCons.copy()] * (self.timeIntegrator.timeOrder+1)
 		solDomain.solInt.solHistPrim = [solDomain.solInt.solPrim.copy()] * (self.timeIntegrator.timeOrder+1)
 
-
-		# gather solution history from stale solution data
 		if self.adaptiveROM and self.adaptiveROMMethod=='AADEIM':
+			"""Gather the available (pre-computed) conservative variables from an input file"""
+
 			assert (os.path.exists(os.path.join(const.unsteadyOutputDir, self.adaptROMConsSnapFName))), \
 				'Path for stale conservative state : '+ str(os.path.join(const.unsteadyOutputDir, self.adaptROMConsSnapFName))+ ' does not exist'
 
@@ -165,7 +164,7 @@ class romDomain:
 			self.timeIntegrator.staleStatetimeOrder = min(self.timeIntegrator.timeOrder,self.staleConsSnapshots.shape[-1])
 			solDomain.timeIntegrator.staleStatetimeOrder = self.timeIntegrator.staleStatetimeOrder
 
[…]
