### 2021-07-29 #3428 @cwentland0
@kyleniemeyer A question, since I'm unfamiliar with software journal reviews in general, especially so when I first submitted this back in April. There have been some significant changes to the code in a couple branches that will soon be merged, including general expansions of capabilities and some class names which were explicitly mentioned in the paper. Should I update the paper to reflect these changes once they're merged (or at the very least, remove any specific references to class names, which was a naive mistake on my part)? I understand the reviewers may have already looked over the paper, but I figure these changes would be mostly superficial, and would not change the bulk of its content. I know there's usually some back and forth with editing, I just don't want to waste any of the reviewers' time.
