### 2021-07-29 #3428 @kyleniemeyer
@cwentland0 good question - this is always tough when software continues to evolve while under review. Minor edits to the paper are fine, especially if you summarize the changes in here for us.

Larger edits to the software, that change capabilities and would essentially require a re-review, should probably wait, so that the version mentioned with a JOSS article is actually the version reviewed.
