### 2021-07-29 #3428 @cwentland0
Maybe "significant changes" was a little overblown on my part. To be more specific, the changes will add two more classes of ROM methods, which would amount to adding two more sub-bullet points under the current "ROMs" bullet point on page 3. Would you consider that a larger edit?
