### 2021-08-02 7369123
hot restart from init_file, fixed error related to density and enthalpy derivative calcs
- can now restart from init_file with 2+ snapshots to preserve time accuracy
- error where calc_sol_jacob was called before calc_rhs_jacob
    - density and enthalpy derivatives were not updated as a result
    - just put calc_sol_jacob before calc_rhs_jacob
