### 2021-08-02 73646b8
allow non-interactive processing of images when 1=1
