+++ perform/visualization/visualization_group.py
@@ -1,5 +1,9 @@
+import os
 from time import sleep
 
+import matplotlib as mpl
+if os.environ["PLT_USE_AGG"] == "1":
+    mpl.use('Agg')
 import matplotlib.pyplot as plt
 
 from perform.constants import FIG_WIDTH_DEFAULT, FIG_HEIGHT_DEFAULT
