+++ perform/visualization/visualization.py
@@ -1,4 +1,8 @@
+import os
+
 import matplotlib as mpl
+if os.environ["PLT_USE_AGG"] == "1":
+    mpl.use('Agg')
 
 mpl.rc("font", family="serif", size="10")
 mpl.rc("axes", labelsize="x-large")
