+++ perform/visualization/field_plot.py
@@ -2,6 +2,10 @@ import os
 from math import floor, log
 
 import numpy as np
+import matplotlib as mpl
+if os.environ["PLT_USE_AGG"] == "1":
+    mpl.use('Agg')
+
 import matplotlib.pyplot as plt
 
 from perform.visualization.visualization import Visualization
