### 2021-08-03 5d7f0b2
fix issue with PLT_USE_AGG, add intermediate results saving
