+++ perform/visualization/probe_plot.py
@@ -1,8 +1,11 @@
 import os
 
 import matplotlib as mpl
-if os.environ["PLT_USE_AGG"] == "1":
-    mpl.use('Agg')
+try:
+    if os.environ["PLT_USE_AGG"] == "1":
+        mpl.use('Agg')
+except KeyError:
+    pass
 import matplotlib.pyplot as plt
 import numpy as np
 
