### 2021-08-29 #3428 @Himscipy
👋🏼  Hi @cwentland0 , Thank you for submitting the paper+ software to JOSS. The developed package will really a nice addition to the scientific community both for research and academic purpose.
I am currently reviewing your code base, one thing which I caught my attention is that "testing" of any sort is missing in the code base. Tests plays a crucial role for any software package and is one of the essential components in the JOSS submissions acceptance criteria, therefore I would like to recommend that you add these testing procedures to the code base. Thank you.
