### 2021-08-30 #3428 @cwentland0
@Himscipy thank you for taking the time to review the code. You're right, I haven't implemented anything in the way of rigorous unit or regression tests, I've only provided a few pre-set cases to ensure base functionality of the solver and few ROMs. I'll get to work on a better testing framework.
