+++ perform/visualization/field_plot.py
@@ -68,7 +68,7 @@ class FieldPlot(Visualization):
 
         self.num_imgs = int(num_steps / vis_interval)
         if self.num_imgs > 0:
-            self.img_string = "%0" + str(floor(log(self.num_imgs, 10)) + 1) + "d"
+            self.img_string = "%0" + str(floor(log(self.num_imgs, 10))) + "d"
         else:
             self.img_string = None
 
@@ -169,17 +169,17 @@ class FieldPlot(Visualization):
             if spec_idx == sol_domain.gas_model.num_species_full:
                 y_data = sol_int.mass_fracs_full[-1, :]
             else:
-                y_data = sol_int.sol_prim[3 + spec_idx - 1, :]
+                y_data = sol_int.sol_prim[3 + spec_idx, :]
         elif var_str[:15] == "density-species":
             spec_idx = int(var_str[16:])
             if spec_idx == sol_domain.gas_model.num_species_full:
                 y_data = sol_int.sol_cons[0, :] * sol_int.mass_fracs_full[-1, :]
             else:
-                y_data = sol_int.sol_cons[3 + spec_idx - 1, :]
+                y_data = sol_int.sol_cons[3 + spec_idx, :]
         elif var_str[:6] == "source":
             spec_idx = int(var_str[7:])
-            y_data = sol_int.reaction_source[spec_idx - 1, sol_domain.direct_samp_idxs]
-        elif var_str == "heat_release":
+            y_data = sol_int.reaction_source[spec_idx, sol_domain.direct_samp_idxs]
+        elif var_str == "heat-release":
             y_data = sol_int.heat_release[sol_domain.direct_samp_idxs]
         else:
             raise ValueError("Invalid field visualization variable:" + str(var_str))
@@ -202,7 +202,7 @@ class FieldPlot(Visualization):
             NumPy array of the x-coordinates for profile to be visualized.
         """
 
-        if (var_str[:6] == "source") or (var_str == "heat_release"):
+        if (var_str[:6] == "source") or (var_str == "heat-release"):
             x_data = sol_domain.mesh.x_cell[sol_domain.direct_samp_idxs]
         else:
             x_data = sol_domain.mesh.x_cell
