### 2021-08-30 d492394
heat_release -> heat-release in SolutionDomain
