### 2021-09-01 b641179
move utils inside perform so other programs can use them, add calc_hr_from_field
