### 2021-09-03 da5a08e
start of big rework, see details
- to make ROM framework more flexible for non-intrusive methods, had to do a huge rework
- ROM framework now broken into:
    - RomDomain: container class for holding a single RomVariableMapping, RomTimeStepper, RomMethod, and multiple RomModels
    - RomMethod: e.g. Galerkin projection, OpInf, continuous Koopman. Dictates evolution of latent variables, and sets some expectations for RomVariableMapping and RomTimeStepper where appropriate.
    - RomVariableMapping: Maps ROM target variables to primitive/conservative state, where applicable
    - RomTimeStepper:
[…]
