+++ perform/rom/rom_model.py
@@ -2,12 +2,11 @@ import os
 import time
 
 import numpy as np
-from scipy.sparse import vstack
 
 from perform.constants import REAL_TYPE
+from perform.rom.rom_space_mapping.linear_space_mapping import LinearSpaceMapping
 
-
-class RomModel:
+class RomModel():
     """Base class for all ROM models.
 
     The RomModel class provides basic functionality that every ROM model should be equipped with,
@@ -46,7 +45,7 @@ class RomModel:
         cent_prof_prim: NumPy array of primitive variable centering profile, if required by ROM method.
     """
 
-    def __init__(self, model_idx, rom_domain, sol_domain):
+    def __init__(self, model_idx, sol_domain, rom_domain):
 
         self.model_idx = model_idx
         self.latent_dim = rom_domain.latent_dims[self.model_idx]
@@ -55,224 +54,18 @@ class RomModel:
         self.num_cells = sol_domain.mesh.num_cells
         self.sol_shape = (self.num_vars, self.num_cells)
 
-        # Just copy some stuff for less clutter
-        model_dir = rom_domain.model_dir
-        self.target_cons = rom_domain.target_cons
-        self.target_prim = rom_domain.target_prim
+        # initialize space mapping
+        space_mapping = rom_domain.rom_dict["space_mapping"]
+        if space_mapping == "linear":
+            self.space_mapping = LinearSpaceMapping(sol_domain, rom_domain, self)
+        else:
+            raise ValueError("Invalid space_mapping: " + str(space_mapping))
 
+        # internal state and related quantities
+        self.sol = np.zeros(self.sol_shape, dtype=REAL_TYPE)
         self.code = np.zeros(self.latent_dim, dtype=REAL_TYPE)
-        self.d_code = np.zeros(self.latent_dim, dtype=REAL_TYPE)
+        self.d_code = np.zeros(self.latent_dim, dtype=REAL_TYPE)  # TODO: is this used?
 
-        # For implicit solve
-        if rom_domain.has_time_integrator:
-            if rom_domain.time_integrator.time_type == "implicit":
-                self.res = np.zeros(self.latent_dim, dtype=REAL_TYPE)
-
-        # Get normalization profiles, if necessary
-        self.norm_sub_prof_cons = None
-        self.norm_sub_prof_prim = None
-        self.norm_fac_prof_cons = None
-        self.norm_fac_prof_prim = None
-        self.cent_prof_cons = None
-        self.cent_prof_prim = None
-        if rom_domain.has_cons_norm:
-            self.norm_sub_prof_cons = self.load_feature_scaling(
-                os.path.join(model_dir, rom_domain.norm_sub_cons_in[self.model_idx]), default="zeros"
-            )
-
-            self.norm_fac_prof_cons = self.load_feature_scaling(
-                os.path.join(model_dir, rom_domain.norm_fac_cons_in[self.model_idx]), default="ones"
-            )
-
-        if rom_domain.has_prim_norm:
-            self.norm_sub_prof_prim = self.load_feature_scaling(
-                os.path.join(model_dir, rom_domain.norm_sub_prim_in[self.model_idx]), default="zeros"
-            )
-
-            self.norm_fac_prof_prim = self.load_feature_scaling(
-                os.path.join(model_dir, rom_domain.norm_fac_prim_in[self.model_idx]), default="ones"
-            )
-
-        # Get centering profiles, if necessary
-        # If cent_ic, just use given initial conditions
-        if rom_domain.has_cons_cent:
-            if rom_domain.cent_ic:
-                self.cent_prof_cons = sol_domain.sol_int.sol_cons[self.var_idxs, :].copy()
-
-            else:
[…]
