### 2021-09-03 fac13ec
re-add linear LSPG
