### 2021-09-07 3ca3c42
very first test, of constants.py
