### 2021-09-07 b055fc6
add some input_funcs tests, first test suite
