### 2021-09-07 bf46d75
testing SystemSolver init, GasModel init and some methods
