### 2021-09-08 74b0166
wrap up GasModel methods tests
