+++ tests/test_suite.py
@@ -5,12 +5,15 @@ import test_input_funcs
 import test_misc_funcs
 import test_mesh
 import test_system_solver
-from test_gas_model import test_gas_model
+from test_gas_model import test_gas_model, test_cpg
 
 
-def suite():
+loader = unittest.TestLoader()
+
+def init_unit_test_suite():
+    """"Test cases for initializing independent classes and unit testing their member methods"""
+
     suite = unittest.TestSuite()
-    loader = unittest.TestLoader()
     suite.addTest(loader.loadTestsFromTestCase(test_constants.ConstantsTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_input_funcs.InputParsersTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_misc_funcs.MiscFuncsTestCase))
@@ -18,9 +21,11 @@ def suite():
     suite.addTests(loader.loadTestsFromTestCase(test_system_solver.SystemSolverInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_gas_model.GasModelInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_gas_model.GasModelMethodsTestCase))
+    suite.addTests(loader.loadTestsFromTestCase(test_cpg.CPGInitTestCase))
+    suite.addTests(loader.loadTestsFromTestCase(test_cpg.GasModelMethodsTestCase))
     return suite
 
 
 if __name__ == "__main__":
     runner = unittest.TextTestRunner(verbosity=3)
-    runner.run(suite())
+    runner.run(init_unit_test_suite())
