### 2021-09-20 6dcbe93
created balanced ROM
