### 2021-09-22 213720b
blackify and flake8, variable and method names do not yet conform to PEP 8
