+++ perform/rom/balanced_rom.py
@@ -1,329 +1,265 @@
-import numpy as np
-import matplotlib.pyplot as plt  
-import copy
-import scipy
-from math import sin, pi, log10
-from numpy.linalg import svd, inv
-from numpy import linalg as LA
-from scipy.sparse.linalg import eigs
 import time
 import os
-import sys
+from math import sin, pi
 
-plt.rc('font', family='serif',size='14')
-plt.rc('axes', labelsize='x-large')
-plt.rc('figure', figsize=(10,10))
+import numpy as np
+from numpy.linalg import svd
+from numpy import linalg as LA
+import matplotlib.pyplot as plt
+
+from perform.input_funcs import read_input_file
+
+plt.rc("font", family="serif", size="14")
+plt.rc("axes", labelsize="x-large")
+plt.rc("figure", figsize=(10, 10))
+
 
 def main():
 
-	rom_input_dir = "../../examples/standing_flame/inputs/brom_inputs"
-	romDict = read_input_file("../../examples/standing_flame/inputs/brom_inputs/brom_params.inp")
+    rom_input_dir = "../../examples/standing_flame/inputs/brom_inputs"
+    romDict = read_input_file("../../examples/standing_flame/inputs/brom_inputs/brom_params.inp")
 
-	BROMOutDir 			= romDict["BROMOutDir"]
-	num_eqs 			= romDict["num_eqs"]
-	nMarkov				= romDict["nMarkov"]
-	plotHSVs			= romDict["plotHSVs"]
-	plotEigs			= romDict["plotEigs"]
+    BROMOutDir = romDict["BROMOutDir"]
+    num_eqs = romDict["num_eqs"]
+    nMarkov = romDict["nMarkov"]
+    plotHSVs = romDict["plotHSVs"]
+    plotEigs = romDict["plotEigs"]
 
-	nMarkov = nMarkov/2
+    nMarkov = nMarkov / 2
 
-	out_dir = os.path.expanduser(BROMOutDir)
-	if not os.path.isdir(BROMOutDir):
-		os.mkdir(BROMOutDir)
+    if not os.path.isdir(BROMOutDir):
+        os.mkdir(BROMOutDir)
 
-	data_file1 = "FOM_Impulse_Resp.npy"
-	in_file = os.path.join(rom_input_dir, data_file1)
-	impulse_resp = np.load(in_file, allow_pickle=True)
-	data_file2 = "steady_init.npy"
-	in_file = os.path.join(rom_input_dir, data_file2)
-	steady_state = np.load(in_file, allow_pickle=True)
+    data_file1 = "FOM_Impulse_Resp.npy"
+    in_file = os.path.join(rom_input_dir, data_file1)
+    impulse_resp = np.load(in_file, allow_pickle=True)
+    data_file2 = "steady_init.npy"
+    in_file = os.path.join(rom_input_dir, data_file2)
+    steady_state = np.load(in_file, allow_pickle=True)
 
-	for varId in range(num_eqs):
+    for varId in range(num_eqs):
 
-		t0 = time.time()
-				
-		Ctrb = impulse_resp[:, varId, 2:] 
-		centProf = np.mean(Ctrb[:, :2*nMarkov], axis=1, keepdims=True)
-		if plotHSVs:
-			sigma = HankelSV(Ctrb-centProf, nMarkov)
-			if (varId == 0): sigmap = sigma
-			if (varId == 1): sigmau = sigma
-			if (varId == 2): sigmaT = sigma
-			if (varId == 3): sigmaY = sigma
-		Ctrb, normFacProf = normalizeBT(Ctrb[:,:2*nMarkov], steady_state[:,varId])
-			
[…]
