+++ tests/test_gas_model/test_gas_model.py
@@ -37,14 +37,14 @@ class GasModelInitTestCase(unittest.TestCase):
         self.assertAlmostEqual(gas.mw_inv_diffs[1], 0.0107142857)
         self.assertEqual(gas.num_eqs, 5)
 
-        mix_mass_matrix = np.array([[1.0, 0.9671682, 1.057371],
-                                    [1.0339463, 1.0, 1.0932651],
-                                    [0.9457416, 0.9146912, 1.0]])
+        mix_mass_matrix = np.array(
+            [[1.0, 0.9671682, 1.057371], [1.0339463, 1.0, 1.0932651], [0.9457416, 0.9146912, 1.0]]
+        )
         self.assertTrue(np.allclose(gas.mix_mass_matrix, mix_mass_matrix))
 
-        mix_inv_mass_matrix = np.array([[0.25, 0.2415229, 0.2635231],
-                                        [0.2581989, 0.25, 0.27116307],
-                                        [0.2357023, 0.2268713, 0.25]])
+        mix_inv_mass_matrix = np.array(
+            [[0.25, 0.2415229, 0.2635231], [0.2581989, 0.25, 0.27116307], [0.2357023, 0.2268713, 0.25]]
+        )
         self.assertTrue(np.allclose(gas.mix_inv_mass_matrix, mix_inv_mass_matrix))
 
 
@@ -68,11 +68,19 @@ class GasModelMethodsTestCase(unittest.TestCase):
 
         # check with all mass fractions
         mass_fracs_out = self.gas.get_mass_frac_array(mass_fracs_in=self.mass_fracs)
-        self.assertTrue(np.array_equal(mass_fracs_out, np.repeat(np.array([[0.231537], [0.755063]], dtype=REAL_TYPE), repeats=2, axis=1)))
+        self.assertTrue(
+            np.array_equal(
+                mass_fracs_out, np.repeat(np.array([[0.231537], [0.755063]], dtype=REAL_TYPE), repeats=2, axis=1)
+            )
+        )
 
         # check with N - 1 mass fractions
         mass_fracs_out = self.gas.get_mass_frac_array(mass_fracs_in=self.mass_fracs[:-1, :])
-        self.assertTrue(np.array_equal(mass_fracs_out, np.repeat(np.array([[0.231537], [0.755063]], dtype=REAL_TYPE), repeats=2, axis=1)))
+        self.assertTrue(
+            np.array_equal(
+                mass_fracs_out, np.repeat(np.array([[0.231537], [0.755063]], dtype=REAL_TYPE), repeats=2, axis=1)
+            )
+        )
 
     def test_calc_all_mass_fracs(self):
 
@@ -87,12 +95,20 @@ class GasModelMethodsTestCase(unittest.TestCase):
     def test_calc_mix_mol_weight(self):
 
         mix_mol_weight = self.gas.calc_mix_mol_weight(self.mass_fracs)
-        self.assertTrue(np.allclose(mix_mol_weight, np.repeat(np.array([[28.9543985]], dtype=REAL_TYPE), repeats=2, axis=1)))
+        self.assertTrue(
+            np.allclose(mix_mol_weight, np.repeat(np.array([[28.9543985]], dtype=REAL_TYPE), repeats=2, axis=1))
+        )
 
     def test_calc_all_mole_fracs(self):
 
         mole_fracs = self.gas.calc_all_mole_fracs(self.mass_fracs)
-        self.assertTrue(np.allclose(mole_fracs, np.repeat(np.array([[0.20950046], [0.78079982], [0.00969972]], dtype=REAL_TYPE), repeats=2, axis=1)))
+        self.assertTrue(
+            np.allclose(
+                mole_fracs,
[…]
