### 2021-09-28 f8a8f23
add constants.py for tests to reuse certain objects more easily
