+++ tests/test_suite.py
@@ -6,12 +6,12 @@ import test_misc_funcs
 import test_mesh
 import test_system_solver
 from test_gas_model import test_gas_model, test_cpg
-
+from test_time_integrator import test_time_integrator, test_explicit_integrator
 
 loader = unittest.TestLoader()
 
 def init_unit_test_suite():
-    """"Test cases for initializing independent classes and unit testing their member methods"""
+    """Test cases for initializing independent classes and unit testing their member methods"""
 
     suite = unittest.TestSuite()
     suite.addTest(loader.loadTestsFromTestCase(test_constants.ConstantsTestCase))
@@ -23,6 +23,11 @@ def init_unit_test_suite():
     suite.addTests(loader.loadTestsFromTestCase(test_gas_model.GasModelMethodsTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_cpg.CPGInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_cpg.GasModelMethodsTestCase))
+    suite.addTests(loader.loadTestsFromTestCase(test_time_integrator.TimeIntegratorTestCase))
+    suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ExplicitTimeIntInitTestCase))
+    suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ClassicRK4InitTestCase))
+    suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.SSPRK3InitTestCase))
+    suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ClassicRK4MethodsTestCase))
     return suite
 
 
