### 2021-11-16 a3b5998
cleanup, improve input path parameter management, calling driver from program
