+++ setup.py
@@ -15,7 +15,7 @@ setup(
     description="One-dimension reacting flow for ROM prototyping",
     long_description=readme,
     license=license,
-    install_requires=["numpy>=1.16.6", "scipy>=1.1.0", "matplotlib>=2.1.0"],
+    install_requires=["numpy>=1.16.6", "scipy>=1.1.0", "matplotlib>=2.1.0", "packaging", "gdown"],
     entry_points={"console_scripts": ["perform = perform.driver:main"]},
     python_requires=">=3.6",
 )
