+++ perform/rom/projection_rom/autoencoder_proj_rom/autoencoder_proj_rom.py
@@ -178,7 +178,7 @@ class AutoencoderProjROM(ProjectionROM):
         return code
 
     def models_init(self, sol_domain, rom_domain):
-        
+
         assert rom_domain.num_models == 1, "Separate models not fixed for non-linear ROMs"
 
     def encode_sol(self, sol_in):
