+++ perform/visualization/probe_plot.py
@@ -1,9 +1,10 @@
 import os
 
 import matplotlib as mpl
+
 try:
     if os.environ["PLT_USE_AGG"] == "1":
-        mpl.use('Agg')
+        mpl.use("Agg")
 except KeyError:
     pass
 import matplotlib.pyplot as plt
@@ -67,7 +68,9 @@ class ProbePlot(Visualization):
 
         self.probe_num = probe_num
         self.probe_vars = probe_vars
-        assert self.probe_num >= 0, "Must provide non-negative (>= 0) integer probe number for probe plot " + str(self.vis_id)
+        assert self.probe_num >= 0, "Must provide non-negative (>= 0) integer probe number for probe plot " + str(
+            self.vis_id
+        )
         assert self.probe_num < num_probes, "probe_num_" + str(self.vis_id) + " must correspond to a valid probe"
         assert vis_vars[0] is not None, "Must provide vis_vars for probe plot " + str(self.vis_id)
 
