### 2021-11-16 9943e7d
add example FOM regression tests
