+++ tests/regression_tests/examples_fom/test_examples_fom.py
@@ -1,0 +1,139 @@
+import unittest
+import subprocess
+import os
+
+import numpy as np
+
+from perform.driver import driver
+
+# TODO: turn off visualization when running these tests
+
+
+class FOMRegressionTests(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        cls.localdir = os.path.dirname(__file__)
+
+        # shock tube
+        cls.shock_tube_test_dir = os.path.join(cls.localdir, "shock_tube")
+        cls.shock_tube_script = os.path.join(cls.shock_tube_test_dir, "get_results.sh")
+        cls.shock_tube_work_dir = os.path.join(cls.localdir, "../../../examples/shock_tube/")  # TODO: this is pretty jank
+        subprocess.call(cls.shock_tube_script)
+
+        # contact surface
+        cls.contact_surface_test_dir = os.path.join(cls.localdir, "contact_surface")
+        cls.contact_surface_script = os.path.join(cls.contact_surface_test_dir, "get_results.sh")
+        cls.contact_surface_work_dir = os.path.join(cls.localdir, "../../../examples/contact_surface/")
+        subprocess.call(cls.contact_surface_script)
+
+        # standing flame
+        cls.standing_flame_test_dir = os.path.join(cls.localdir, "standing_flame")
+        cls.standing_flame_script = os.path.join(cls.standing_flame_test_dir, "get_results.sh")
+        cls.standing_flame_work_dir = os.path.join(cls.localdir, "../../../examples/standing_flame/")
+        subprocess.call(cls.standing_flame_script)
+
+        # transient flame
+        cls.transient_flame_test_dir = os.path.join(cls.localdir, "transient_flame")
+        cls.transient_flame_script = os.path.join(cls.transient_flame_test_dir, "get_results.sh")
+        cls.transient_flame_work_dir = os.path.join(cls.localdir, "../../../examples/transient_flame/")
+        subprocess.call(cls.transient_flame_script)
+
+    @classmethod
+    def tearDownClass(cls):
+        subprocess.call("rm " + cls.shock_tube_test_dir + "/*.npy", shell=True)
+        subprocess.call("rm " + cls.contact_surface_test_dir + "/*.npy", shell=True)
+        subprocess.call("rm " + cls.standing_flame_test_dir + "/*.npy", shell=True)
+        subprocess.call("rm " + cls.transient_flame_test_dir + "/*.npy", shell=True)
+
+    def test_shock_tube(self):
+        
+        # run shock tube case and load results
+        driver(self.shock_tube_work_dir)
+        unst_dir = os.path.join(self.shock_tube_work_dir, "unsteady_field_results")
+        probe_dir = os.path.join(self.shock_tube_work_dir, "probe_results")
+        sol_cons_test = np.load(os.path.join(unst_dir, "sol_cons_FOM.npy"))
+        sol_prim_test = np.load(os.path.join(unst_dir, "sol_prim_FOM.npy"))
+        probe_1_test = np.load(os.path.join(probe_dir, "probe_pressure_velocity_density_temperature_1_FOM.npy"))
+        probe_2_test = np.load(os.path.join(probe_dir, "probe_pressure_velocity_density_temperature_2_FOM.npy"))
+
+        # load truth results
+        sol_cons_truth = np.load(os.path.join(self.shock_tube_test_dir, "sol_cons_FOM.npy"))
+        sol_prim_truth = np.load(os.path.join(self.shock_tube_test_dir, "sol_prim_FOM.npy"))
[…]
