+++ .github/workflows/perform_basic.yml
@@ -1,0 +1,56 @@
+name: Basic PERFORM actions
+
+# Runs unit and integration tests, and checks code formatting
+# Executed every time code is pushed to any branch, on any pull request, and at 5:00 UTC
+
+on: 
+  push:
+  pull_request:
+  # schedule:
+  #   - cron: 0 5 * * *
+
+
+jobs:
+
+  # Run unit and integration tests, which should be pretty quick
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Set up Python 3.8
+        uses: actions/setup-python@v2
+        with:
+          python-version: '3.8'
+
+      - name: Install PERFORM and dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install .
+
+      - name: Run unit tests
+        run: |
+          cd test/unit_tests
+          python test_suite.py
+
+  # Check code formatting using black
+  # black:
+  #   runs-on: ubuntu-latest
+
+  #   steps:
+  #   - uses: actions/checkout@v2
+
+  #   - name: Set up Python 3.8
+  #     uses: actions/setup-python@v2
+  #     with:
+  #       python-version: '3.8'
+
+  #   - name: Check black
+  #     run: |
+  #       pip install black==21.10b0
+  #       black perform --check --diff -l 120 --target-version py38
+
+  # Check consistency with PEP8 using flake8
+  # .flake8 
+  # flake8:
