+++ .github/workflows/perform_basic.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Install PERFORM and dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install .
+          pip install -e .
 
       - name: Run unit tests
         run: |
