+++ perform/rom/ml_library/pytorch_library.py
@@ -2,8 +2,7 @@ from perform.rom.ml_library.ml_library import MLLibrary
 
 
 class PyTorchLibrary(MLLibrary):
-    """Class for accessing PyTorch functionalities.
-    """
+    """Class for accessing PyTorch functionalities."""
 
     def __init__(self, rom_domain):
         raise ValueError("PyTorch MLLibrary not implemented yet")
