+++ perform/rom/nonintrusive_rom/autoencoder_rom/autoencoder_koopman/pan2020.py
@@ -2,7 +2,4 @@ from perform.rom.nonintrusive_rom.autoencoder_rom.autoencoder_koopman.autoencode
 
 
 class AEKoopmanPan2020(AutoencoderKoopman):
-    """Class implementing the continuous-time variant of the autoencoder Koopman method by Pan (2020).
-
-
-    """
+    """Class implementing the continuous-time variant of the autoencoder Koopman method by Pan (2020)."""
