### 2021-11-17 2c43895
add flake8 to GHA
