### 2021-11-17 d2bbd9e
correct flake8 errors
