### 2021-11-17 c66f704
add nightly schedule for basic GHA
