+++ .github/workflows/perform_intesive.yml
@@ -1,0 +1,37 @@
+name: Time-intensive PERFORM actions
+
+# Runs regression tests
+# Executed on pull request to main, and at 5:00 UTC
+
+on:
+  push:
+  # pull_request:
+  #   branches:
+  #     - main
+  # schedule:
+  #   - cron: 0 5 * * *
+
+
+jobs:
+
+  # Run regression tests, which take a bit longer
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Set up Python 3.8
+        uses: actions/setup-python@v2
+        with:
+          python-version: '3.8'
+
+      - name: Install PERFORM and dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .
+
+      - name: Run unit tests
+        run: |
+          cd tests/regression_tests
+          python test_suite.py
