+++ tests/unit_tests/test_suite.py
@@ -12,8 +12,8 @@ from test_time_integrator import test_time_integrator, test_explicit_integrator
 loader = unittest.TestLoader()
 
 
-def init_unit_test_suite():
-    """Test cases for initializing independent classes and unit testing their member methods"""
+def indep_unit_test_suite():
+    """Test cases for independent class unit tests"""
 
     suite = unittest.TestSuite()
     suite.addTest(loader.loadTestsFromTestCase(test_constants.ConstantsTestCase))
@@ -21,10 +21,24 @@ def init_unit_test_suite():
     suite.addTests(loader.loadTestsFromTestCase(test_misc_funcs.MiscFuncsTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_mesh.MeshTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_system_solver.SystemSolverInitTestCase))
+    return suite
+
+
+def gas_model_test_suite():
+    """Test cases for gas model unit tests"""
+
+    suite = unittest.TestSuite()
     suite.addTests(loader.loadTestsFromTestCase(test_gas_model.GasModelInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_gas_model.GasModelMethodsTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_cpg.CPGInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_cpg.GasModelMethodsTestCase))
+    return suite
+
+
+def time_int_test_suite():
+    """Test cases for time integrator unit tests"""
+
+    suite = unittest.TestSuite()
     suite.addTests(loader.loadTestsFromTestCase(test_time_integrator.TimeIntegratorTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ExplicitTimeIntInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ClassicRK4InitTestCase))
@@ -32,7 +46,8 @@ def init_unit_test_suite():
     suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ClassicRK4MethodsTestCase))
     return suite
 
-
 if __name__ == "__main__":
     runner = unittest.TextTestRunner(verbosity=3)
-    runner.run(init_unit_test_suite())
+    runner.run(indep_unit_test_suite())
+    runner.run(gas_model_test_suite())
+    runner.run(time_int_test_suite())
