+++ tests/unit_tests/test_time_integrator/test_explicit_integrator.py
@@ -74,19 +74,41 @@ class ClassicRK4MethodsTestCase(unittest.TestCase):
 
         # set up time integrator
         self.time_int = ClassicRK4(self.param_dict)
-        self.time_int.subiter = 0
 
         self.rhs = np.array([[0, 3], [1, 4], [2, 5]])
+        self.rk_rhs = [
+            np.array([[0, 3], [1, 4], [2, 5]]),
+            np.array([[1, 4], [2, 5], [3, 6]]),
+            np.array([[2, 5], [3, 6], [3, 7]]),
+            np.array([[3, 6], [4, 7], [4, 8]]),
+        ]
 
     def test_solve_sol_change_subiter(self):
 
         self.time_int.rk_rhs[0] = self.rhs.copy()
         dsol = self.time_int.solve_sol_change_subiter(self.rhs)
-        self.assertTrue(np.array_equal(dsol, np.array([[0, 1.5], [0.5, 2], [1, 2.5]])))
+        self.assertTrue(np.allclose(dsol, np.array([[0, 1.5], [0.5, 2], [1, 2.5]])))
 
-    # def test_solve_sol_change_iter(self):
+    def test_solve_sol_change_iter(self):
 
-    # def test_solve_sol_change(self):
+        self.time_int.rk_rhs = self.rk_rhs
+        dsol = self.time_int.solve_sol_change_iter(self.rhs)
+        self.assertTrue(np.allclose(dsol, np.array([[1.5, 4.5], [2.5, 5.5], [3, 6.5]])))
+
+    def test_solve_sol_change(self):
+
+        for self.time_int.subiter in range(self.time_int.subiter_max):
+
+            dsol = self.time_int.solve_sol_change(self.rk_rhs[self.time_int.subiter])
+
+            if self.time_int.subiter == 0:
+                self.assertTrue(np.allclose(dsol, np.array([[0, 1.5e-7], [5e-8, 2e-7], [1e-7, 2.5e-7]])))
+            elif self.time_int.subiter == 1:
+                self.assertTrue(np.allclose(dsol, np.array([[5.0e-8, 2.0e-7], [1.0e-7, 2.5e-7], [1.5e-07, 3.0e-7]])))
+            elif self.time_int.subiter == 2:
+                self.assertTrue(np.allclose(dsol, np.array([[2.0e-7, 5.e-7], [3.0e-7, 6.0e-7], [3.0e-7, 7.0e-7]])))
+            elif self.time_int.subiter == 3:
+                self.assertTrue(np.allclose(dsol, np.array([[1.5e-7, 4.5e-7], [2.5e-7, 5.5e-7], [3e-7, 6.5e-7]])))
 
 
 class SSPRK3InitTestCase(unittest.TestCase):
@@ -107,5 +129,48 @@ class SSPRK3InitTestCase(unittest.TestCase):
         self.assertEqual(time_int.rk_b_vals.shape, (3,))
         self.assertEqual(time_int.rk_c_vals.shape, (3,))
 
+
+class SSPRK3MethodsTestCase(unittest.TestCase):
+    def setUp(self):
+        # set up param_dict
+        self.param_dict = {}
+        self.param_dict["dt"] = 1e-7
+        self.param_dict["time_scheme"] = "ssp_rk3"
+        self.param_dict["time_order"] = 3
+
+        # set up time integrator
+        self.time_int = SSPRK3(self.param_dict)
+
+        self.rhs = np.array([[0, 3], [1, 4], [2, 5]])
+        self.rk_rhs = [
+            np.array([[0, 3], [1, 4], [2, 5]]),
+            np.array([[1, 4], [2, 5], [3, 6]]),
[…]
