+++ tests/unit_tests/test_suite.py
@@ -6,7 +6,7 @@ import test_misc_funcs
 import test_mesh
 import test_system_solver
 from test_gas_model import test_gas_model, test_cpg
-from test_time_integrator import test_time_integrator, test_explicit_integrator
+from test_time_integrator import test_time_integrator, test_explicit_integrator, test_implicit_integrator
 
 
 loader = unittest.TestLoader()
@@ -16,11 +16,13 @@ def indep_unit_test_suite():
     """Test cases for independent class unit tests"""
 
     suite = unittest.TestSuite()
+    
     suite.addTest(loader.loadTestsFromTestCase(test_constants.ConstantsTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_input_funcs.InputParsersTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_misc_funcs.MiscFuncsTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_mesh.MeshTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_system_solver.SystemSolverInitTestCase))
+    
     return suite
 
 
@@ -28,10 +30,12 @@ def gas_model_test_suite():
     """Test cases for gas model unit tests"""
 
     suite = unittest.TestSuite()
+    
     suite.addTests(loader.loadTestsFromTestCase(test_gas_model.GasModelInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_gas_model.GasModelMethodsTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_cpg.CPGInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_cpg.GasModelMethodsTestCase))
+    
     return suite
 
 
@@ -39,12 +43,18 @@ def time_int_test_suite():
     """Test cases for time integrator unit tests"""
 
     suite = unittest.TestSuite()
+    
     suite.addTests(loader.loadTestsFromTestCase(test_time_integrator.TimeIntegratorTestCase))
+    
     suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ExplicitTimeIntInitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ClassicRK4InitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.SSPRK3InitTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.ClassicRK4MethodsTestCase))
     suite.addTests(loader.loadTestsFromTestCase(test_explicit_integrator.SSPRK3MethodsTestCase))
+
+    suite.addTests(loader.loadTestsFromTestCase(test_implicit_integrator.ImplicitTimeIntInitTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_implicit_integrator.BDFTestCase))
+
     return suite
 
 if __name__ == "__main__":
