+++ .github/workflows/perform_intesive.yml
@@ -8,7 +8,7 @@ on:
     branches:
       - main
   schedule:
-    - cron: 0 5 * * *
+    - cron: "10 * * * *"
 
 
 jobs:
