### 2021-11-17 3d1af99
switch back intensive GHA to 5:00, won't run until it's merged to main anyway
