+++ .github/workflows/perform_intesive.yml
@@ -8,7 +8,7 @@ on:
     branches:
       - main
   schedule:
-    - cron: "10 * * * *"
+    - cron: "0 5 * * *"
 
 
 jobs:
