### 2021-11-17 2392d71
add first stab at SolutionPhys tests, add integration tests to GHA
