### 2021-11-17 f4c61cb
start SolutionInterior testing
