+++ tests/integration_tests/test_suite.py
@@ -1,16 +1,22 @@
 import unittest
 
-from test_solution import test_solution_phys
+from test_solution import test_solution_phys, test_solution_interior
 
 loader = unittest.TestLoader()
 
 
 def solution_integration_test_suite():
-    """Test cases for independent class unit tests"""
 
     suite = unittest.TestSuite()
+
+    # initializations
     suite.addTests(loader.loadTestsFromTestCase(test_solution_phys.SolutionPhysInitTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_solution_interior.SolutionIntInitTestCase))
+
+    # methods
     suite.addTest(loader.loadTestsFromTestCase(test_solution_phys.SolutionPhysMethodsTestCase))
+    
+
     return suite
 
 
