### 2021-11-19 83061b8
convert test_solution_phys to .npy checks, continue test_solution_interior
