+++ tests/integration_tests/test_solution/test_solution_phys.py
@@ -1,4 +1,5 @@
 import unittest
+import os
 
 import numpy as np
 
@@ -9,6 +10,9 @@ from perform.solution.solution_phys import SolutionPhys
 
 class SolutionPhysInitTestCase(unittest.TestCase):
     def setUp(self):
+
+        self.output_mode = bool(os.environ["PERFORM_TEST_OUTPUT_MODE"])
+        self.output_dir = os.environ["PERFORM_TEST_OUTPUT_DIR"]
 
         self.chem_dict = CHEM_DICT_AIR
         self.gas = CaloricallyPerfectGas(self.chem_dict)
@@ -37,36 +41,42 @@ class SolutionPhysInitTestCase(unittest.TestCase):
 
         # check initialization from primitive state
         sol = SolutionPhys(self.gas, self.num_cells, sol_prim_in=self.sol_prim_in)
-        self.assertTrue(np.array_equal(sol.sol_prim, self.sol_prim_in))
-        self.assertTrue(np.allclose(
-            sol.sol_cons,
-            np.array([
-                [11.8162321, 0.910169230],
-                [23.6324642, 0.910169230],
-                [2.51369841e6, 2.51981100e5],
-                [4.72649284, 0.546101538],
-                [7.08973927, 0.364067692],
-            ])
-        ))
 
-        # TODO: a LOT of checking of other variables
+        if self.output_mode:
+
+            np.save(os.path.join(self.output_dir, "sol_phys_init_sol_cons.npy"), sol.sol_cons)
+
+        else:
+
+            self.assertTrue(np.array_equal(sol.sol_prim, self.sol_prim_in))
+            self.assertTrue(np.allclose(
+                sol.sol_cons,
+                np.load(os.path.join(self.output_dir, "sol_phys_init_sol_cons.npy"))
+            ))
+
+            # TODO: a LOT of checking of other variables
 
         # check initialization from conservative state (just check that primitive state is correct this time)
         sol = SolutionPhys(self.gas, self.num_cells, sol_cons_in=self.sol_cons_in)
-        self.assertTrue(np.array_equal(sol.sol_cons, self.sol_cons_in))
-        self.assertTrue(np.allclose(
-            sol.sol_prim,
-            np.array([
-                [9.74659260e5, 9.72295116e4],
-                [2, 1],
-                [292.800001, 393.312487],
-                [0.4, 0.6],
-                [0.6, 0.4],
-            ])
-        ))
+
+        if self.output_mode:
+
+            np.save(os.path.join(self.output_dir, "sol_phys_init_sol_prim.npy"), sol.sol_prim)
+
+        else:
+
+            self.assertTrue(np.array_equal(sol.sol_cons, self.sol_cons_in))
+            self.assertTrue(np.allclose(
+                sol.sol_prim,
+                np.load(os.path.join(self.output_dir, "sol_phys_init_sol_prim.npy"))
+            ))
 
 class SolutionPhysMethodsTestCase(unittest.TestCase):
+
     def setUp(self):
+
+        self.output_mode = bool(os.environ["PERFORM_TEST_OUTPUT_MODE"])
+        self.output_dir = os.environ["PERFORM_TEST_OUTPUT_DIR"]
 
         self.chem_dict = CHEM_DICT_AIR
         self.gas = CaloricallyPerfectGas(self.chem_dict)
@@ -87,61 +97,67 @@ class SolutionPhysMethodsTestCase(unittest.TestCase):
 
         self.sol.update_density_enthalpy_derivs()
 
-        # check density derivatives
-        self.assertTrue(np.allclose(
-            self.sol.d_rho_d_press,
-            np.array([1.18162321e-5, 9.10169230e-6]),
-        ))
-        self.assertTrue(np.allclose(
[…]
