### 2021-11-19 8d2b807
convert test_solution_interior to .npy
