+++ .github/workflows/perform_basic.yml
@@ -37,6 +37,7 @@ jobs:
       - name: Run integration tests
         run: |
           cd tests/integration_tests
+          ls
           python test_suite.py
 
   # Check code formatting using black
