### 2021-11-19 fc61c32
check if subprocess needs ./ for current directory references
