### 2021-11-19 f9161c8
remove debugging print statements
