### 2021-11-19 bc8bdbc
add test_finite_rate_irrev_reaction
