### 2021-11-19 771cef4
correct error in .npy check, add test_invisc_flux
