+++ tests/integration_tests/test_suite.py
@@ -6,6 +6,7 @@ from argparse import ArgumentParser
 
 from test_solution import test_solution_phys, test_solution_interior
 from test_reaction_model import test_finite_rate_irrev_reaction
+from test_flux.invisc_flux import test_invisc_flux
 
 loader = unittest.TestLoader()
 
@@ -23,6 +24,7 @@ def integration_test_suite(output_mode=False):
     suite.addTest(loader.loadTestsFromTestCase(test_solution_phys.SolutionPhysMethodsTestCase))
     suite.addTest(loader.loadTestsFromTestCase(test_solution_interior.SolutionIntMethodsTestCase))
     suite.addTest(loader.loadTestsFromTestCase(test_finite_rate_irrev_reaction.FiniteRateIrrevReactionMethodsTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_invisc_flux.InviscFluxMethodsTestCase))
 
     return suite
 
