### 2021-11-19 190418a
start test_solution_domain, add test_solution_boundary
