+++ tests/integration_tests/test_solution/test_solution_boundary/test_solution_boundary.py
@@ -1,0 +1,139 @@
+import unittest
+import os
+import shutil
+
+import numpy as np
+
+from constants import CHEM_DICT_REACT
+import perform.constants as constants
+from perform.system_solver import SystemSolver
+from perform.gas_model.calorically_perfect_gas import CaloricallyPerfectGas
+from perform.solution.solution_boundary.solution_boundary import SolutionBoundary
+
+class SolutionBoundaryInitTestCase(unittest.TestCase):
+    def setUp(self):
+        # set chemistry
+        self.chem_dict = CHEM_DICT_REACT
+        self.gas = CaloricallyPerfectGas(self.chem_dict)
+
+        # generate working directory
+        self.test_dir = "test_dir"
+        if os.path.isdir(self.test_dir):
+            shutil.rmtree(self.test_dir)
+        os.mkdir(self.test_dir)
+
+        self.bound_type = "inlet"
+        self.press = 1e6
+        self.vel = 2.0
+        self.temp = 300.0
+        self.mass_fracs = [0.6, 0.4]
+        self.rho = 11.8
+        self.pert_type = "pressure"
+        self.pert_perc = 0.025
+        self.pert_freq = [125e3]
+
+        # generate file with necessary input files
+        self.test_file = os.path.join(self.test_dir, constants.PARAM_INPUTS)
+        with open(self.test_file, "w") as f:
+            f.write('init_file = "test_init_file.npy"\n')
+            f.write("dt = 1e-8\n")
+            f.write('time_scheme = "bdf"\n')
+            f.write("num_steps = 100\n")
+
+            f.write("press_" + self.bound_type + " = " + str(self.press) + "\n")
+            f.write("vel_" + self.bound_type + " = " + str(self.vel) + "\n")
+            f.write("temp_" + self.bound_type + " = " + str(self.temp) + "\n")
+            f.write("mass_fracs_" + self.bound_type + " = " + str(self.mass_fracs) + "\n")
+            f.write("rho_" + self.bound_type + " = " + str(self.rho) + "\n")
+            f.write("pert_type_" + self.bound_type + " = \"" + str(self.pert_type) + "\"\n")
+            f.write("pert_perc_" + self.bound_type + " = " + str(self.pert_perc) + "\n")
+            f.write("pert_freq_" + self.bound_type + " = " + str(self.pert_freq) + "\n")
+
+        # set SystemSolver
+        self.solver = SystemSolver(self.test_dir)
+
+    def tearDown(self):
+
+        if os.path.isdir(self.test_dir):
+            shutil.rmtree(self.test_dir)
+
+    def test_solution_boundary_init(self):
+
+        sol = SolutionBoundary(self.gas, self.solver, "inlet")
+        
+        self.assertEqual(sol.press, self.press)
+        self.assertEqual(sol.vel, self.vel)
+        self.assertEqual(sol.temp, self.temp)
+        self.assertTrue(np.array_equal(sol.mass_fracs, self.mass_fracs))
+        self.assertEqual(sol.rho, self.rho)
+        self.assertEqual(sol.pert_type, self.pert_type)
+        self.assertEqual(sol.pert_perc, self.pert_perc)
+        self.assertTrue(np.array_equal(sol.pert_freq, self.pert_freq))
+
+
+class SolutionBoundaryMethodsTestCase(unittest.TestCase):
+    def setUp(self):
+
+        self.output_mode = bool(int(os.environ["PERFORM_TEST_OUTPUT_MODE"]))
+        self.output_dir = os.environ["PERFORM_TEST_OUTPUT_DIR"]
+
+        # set chemistry
+        self.chem_dict = CHEM_DICT_REACT
+        self.gas = CaloricallyPerfectGas(self.chem_dict)
+
[…]
