### 2021-11-20 505e0f0
add test_solution_inlet
