### 2021-11-20 768939c
correct some weird string formatting from a long time ago
