### 2021-11-20 32b1126
added test_solution_outlet, corrected error in meanflow outlet
