### 2021-11-29 e93b3c8
add test_solution_domain_init, revert test_suite to online
