+++ tests/integration_tests/test_solution/test_solution_domain.py
@@ -37,6 +37,8 @@ class SolutionDomainInitTestCase(unittest.TestCase):
             for key, item in CHEM_DICT_REACT.items():
                 if isinstance(item, str):
                     f.write(key + " = \"" + str(item) + "\"\n")
+                elif isinstance(item, list) or isinstance(item, np.ndarray):
+                    f.write(key + ' = [' + ','.join(str(val) for val in item) + ']\n')
                 else:
                     f.write(key + " = " + str(item) + "\n")
 
@@ -52,30 +54,28 @@ class SolutionDomainInitTestCase(unittest.TestCase):
         # generate solver input files
         self.inp_file = os.path.join(self.test_dir, constants.PARAM_INPUTS)
         with open(self.inp_file, "w") as f:
-            f.write('chem_file = "' + self.chem_file + '"\n')
-            f.write('mesh_file = "' + self.mesh_file + '"\n')
-            f.write('init_file = "test_init_file.npy"\n')
-            f.write("dt = 1e-7\n")
-            f.write('time_scheme = "bdf"\n')
-            f.write("num_steps = 100\n")
-            f.write("invisc_flux_scheme")
-            f.write("visc_flux_scheme")
-            f.write("space_order")
-            f.write("grad_limiter")
-            f.write("bound_cond_inlet")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
-            f.write("")
+            f.write('chem_file = "./chem.inp" \n')
+            f.write('mesh_file = "./mesh.inp" \n')
+            f.write('init_file = "test_init_file.npy" \n')
+            f.write("dt = 1e-7 \n")
+            f.write('time_scheme = "bdf" \n')
+            f.write("time_order = 2 \n")
+            f.write("num_steps = 10 \n")
+            f.write('invisc_flux_scheme = "roe" \n')
+            f.write('visc_flux_scheme = "standard" \n')
+            f.write("space_order = 2 \n")
+            f.write('grad_limiter = "barth_face" \n')
+            f.write('bound_cond_inlet = "meanflow" \n')
+            f.write("press_inlet = 1003706.0 \n")
+            f.write("temp_inlet = 1000.0 \n")
+            f.write("vel_inlet = 1853.0 \n")
+            f.write("rho_inlet = 3944.0 \n")
+            f.write("mass_fracs_inlet = [0.6, 0.4] \n")
+            f.write('bound_cond_outlet = "meanflow" \n')
+            f.write("press_outlet = 99830.0 \n")
+            f.write("vel_outlet = 169.0 \n")
+            f.write("rho_outlet = 328.0 \n")
+            f.write("mass_fracs_outlet = [0.4, 0.6] \n")
 
         # set SystemSolver and time integrator
         self.solver = SystemSolver(self.test_dir)
@@ -88,4 +88,18 @@ class SolutionDomainInitTestCase(unittest.TestCase):
 
     def test_solution_domain_init(self):
 
-        pass
+        sol = SolutionDomain(self.solver)
+        
+        if self.output_mode:
+
+            np.save(os.path.join(self.output_dir, "sol_domain_int_init_sol_cons.npy"), sol.sol_int.sol_cons)
+
+        else:
+
+            self.assertTrue(np.array_equal(sol.sol_int.sol_prim, self.sol_prim_in))
+            self.assertTrue(np.allclose(
+                sol.sol_int.sol_cons,
+                np.load(os.path.join(self.output_dir, "sol_domain_int_init_sol_cons.npy"))
+            ))
+        
+            # TODO: a LOT of checking of other variables
