### 2021-11-29 c035c41
blackify, adjust inlet/outlet sol_prim_init to work with second-order flux
