### 2021-11-29 232f7b6
attempt to use coveralls-python instead
