+++ .github/workflows/perform_basic.yml
@@ -28,7 +28,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -e .
-          pip install coverage
+          pip install coverage coveralls
 
       - name: Run unit tests
         run: |
@@ -40,17 +40,19 @@ jobs:
           cd tests/
           COVERAGE_FILE=.coverage_int coverage run --source=../perform ./integration_tests/test_suite.py
 
-      - name: Combine coverage reports
+      # - name: Combine coverage reports
+      #   run: |
+      #     cd tests/
+      #     coverage combine .coverage_unit .coverage_int
+
+      - name: Report coverage
+        with:
+          github-token: ${{ secrets.GITHUB_TOKEN }}
         run: |
           cd tests/
           coverage combine .coverage_unit .coverage_int
-
-      - name: Report coverage
-        uses: coverallsapp/github-action@master
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          path-to-lcov: "./tests/.coverage"
-        
+          coveralls --service=github
+          
 
   # Check code formatting using black
   black:
