### 2021-11-29 3ababdc
use env GITHUB_TOKEN instead
