### 2021-11-29 88cbaef
add .coveragerc for some exclusions
