### 2021-11-29 606f6f2
add calc_rhs_jacob and calc_res_jacob
