### 2021-11-29 eef2154
move sample SolutionDomain to constants
