### 2021-11-30 e3c1bbf
start FieldPlot testing
