+++ tests/integration_tests/test_suite.py
@@ -9,6 +9,7 @@ from test_solution.test_solution_boundary import test_solution_boundary, test_so
 from test_reaction_model import test_finite_rate_irrev_reaction
 from test_flux.invisc_flux import test_invisc_flux
 from test_limiter import test_limiters
+from test_visualization import test_field_plot
 
 loader = unittest.TestLoader()
 
@@ -38,6 +39,8 @@ def integration_test_suite(output_mode=False):
 
     # classes that depend on SolutionDomain
     suite.addTest(loader.loadTestsFromTestCase(test_limiters.LimiterMethodsTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_field_plot.FieldPlotInitTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_field_plot.FieldPlotMethodsTestCase))
 
     return suite
 
@@ -48,9 +51,11 @@ if __name__ == "__main__":
     # In output mode, will not check method results, but will instead save them to disk
     # These outputs are retrieved from remote storage and used for comparison when not in output mode
     parser = ArgumentParser()
-    parser.add_argument("-o", action="store_true", dest="output_mode", default=False, help="Run method tests in output mode")
+    parser.add_argument(
+        "-o", action="store_true", dest="output_mode", default=False, help="Run method tests in output mode"
+    )
     output_mode = parser.parse_args().output_mode
-    
+
     # clear out the current output directory and remake it
     localdir = os.path.dirname(__file__)
     # handle weirdness with GHA and subprocess
@@ -62,7 +67,7 @@ if __name__ == "__main__":
         shutil.rmtree(outputdir)
     if not os.path.isdir(outputdir):
         os.mkdir(outputdir)
-    
+
     os.environ["PERFORM_TEST_OUTPUT_DIR"] = outputdir
 
     if output_mode:
@@ -74,4 +79,3 @@ if __name__ == "__main__":
 
     runner = unittest.TextTestRunner(verbosity=3)
     runner.run(integration_test_suite(output_mode=output_mode))
-
