+++ tests/integration_tests/test_solution/test_solution_domain.py
@@ -9,7 +9,6 @@ from perform.system_solver import SystemSolver
 from perform.solution.solution_domain import SolutionDomain
 
 
-
 class SolutionDomainInitTestCase(unittest.TestCase):
     def setUp(self):
 
@@ -26,12 +25,14 @@ class SolutionDomainInitTestCase(unittest.TestCase):
         solution_domain_setup(self.test_dir)
 
         # generate initial condition file
-        self.sol_prim_in = np.array([
-            [1e6, 9e5],
-            [2.0, 1.0],
-            [1000.0, 1200.0],
-            [0.6, 0.4],
-        ])
+        self.sol_prim_in = np.array(
+            [
+                [1e6, 9e5],
+                [2.0, 1.0],
+                [1000.0, 1200.0],
+                [0.6, 0.4],
+            ]
+        )
         np.save(os.path.join(self.test_dir, "test_init_file.npy"), self.sol_prim_in)
 
         # set SystemSolver
@@ -45,7 +46,7 @@ class SolutionDomainInitTestCase(unittest.TestCase):
     def test_solution_domain_init(self):
 
         sol = SolutionDomain(self.solver)
-        
+
         if self.output_mode:
 
             np.save(os.path.join(self.output_dir, "sol_domain_int_init_sol_cons.npy"), sol.sol_int.sol_cons)
@@ -53,11 +54,12 @@ class SolutionDomainInitTestCase(unittest.TestCase):
         else:
 
             self.assertTrue(np.array_equal(sol.sol_int.sol_prim, self.sol_prim_in))
-            self.assertTrue(np.allclose(
-                sol.sol_int.sol_cons,
-                np.load(os.path.join(self.output_dir, "sol_domain_int_init_sol_cons.npy"))
-            ))
-        
+            self.assertTrue(
+                np.allclose(
+                    sol.sol_int.sol_cons, np.load(os.path.join(self.output_dir, "sol_domain_int_init_sol_cons.npy"))
+                )
+            )
+
             # TODO: a LOT of checking of other variables
 
 
@@ -77,12 +79,14 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
         solution_domain_setup(self.test_dir)
 
         # generate initial condition file
-        self.sol_prim_in = np.array([
-            [1e6, 9e5],
-            [2.0, 1.0],
-            [1000.0, 1200.0],
-            [0.6, 0.4],
-        ])
+        self.sol_prim_in = np.array(
+            [
+                [1e6, 9e5],
+                [2.0, 1.0],
+                [1000.0, 1200.0],
+                [0.6, 0.4],
+            ]
+        )
         np.save(os.path.join(self.test_dir, "test_init_file.npy"), self.sol_prim_in)
 
         # set SystemSolver and SolutionDomain
@@ -104,10 +108,9 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
 
         else:
 
-            self.assertTrue(np.allclose(
-                self.sol_domain.sol_int.rhs,
-                np.load(os.path.join(self.output_dir, "sol_domain_rhs.npy"))
-            ))
+            self.assertTrue(
+                np.allclose(self.sol_domain.sol_int.rhs, np.load(os.path.join(self.output_dir, "sol_domain_rhs.npy")))
+            )
 
     def test_calc_rhs_jacob(self):
 
@@ -122,18 +125,15 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
 
         else:
 
-            self.assertTrue(np.allclose(
-                rhs_jacob_center,
-                np.load(os.path.join(self.output_dir, "sol_domain_rhs_jacob_center.npy"))
-            ))
-            self.assertTrue(np.allclose(
-                rhs_jacob_left,
[…]
