+++ tests/integration_tests/constants.py
@@ -91,10 +91,14 @@ def solution_domain_setup(run_dir):
         f.write("rho_outlet = 2958.0 \n")
         f.write("mass_fracs_outlet = [0.4, 0.6] \n")
         f.write("probe_locs = [-1.0, 5e-6, 1.0] \n")
-        f.write('probe_vars = ["pressure", "velocity"] \n')
+        f.write('probe_vars = ["pressure", "temperature", "species-0", "density"] \n')
         f.write("vis_show = False \n")
         f.write("vis_save = True \n")
         f.write("vis_interval = 1 \n")
         f.write('vis_type_0 = "field" \n')
         f.write('vis_var_0 = ["temperature", "density", "pressure", "species-0"] \n')
         f.write("vis_y_bounds_0 = [[500, 1500], [1.8, 2.6], [1.2e6, 8e5], [-0.1, 1.1]] \n")
+        f.write('vis_type_1 = "probe" \n')
+        f.write('vis_var_1 = ["temperature", "density", "pressure", "species-0"] \n')
+        f.write("vis_y_bounds_1 = [[500, 1500], [1.8, 2.6], [1.2e6, 8e5], [-0.1, 1.1]] \n")
+        f.write("probe_num_1 = 1")
