+++ tests/integration_tests/test_visualization/test_probe_plot.py
@@ -5,6 +5,7 @@ import shutil
 import numpy as np
 
 from constants import solution_domain_setup
+from perform.constants import REAL_TYPE
 from perform.input_funcs import catch_list
 from perform.system_solver import SystemSolver
 from perform.solution.solution_domain import SolutionDomain
@@ -72,3 +73,118 @@ class ProbePlotInitTestCase(unittest.TestCase):
         for bound_idx, bound in enumerate(self.vis_x_bounds):
             self.assertTrue(np.array_equal(bound, plot.vis_x_bounds[bound_idx]))
             self.assertTrue(np.array_equal(self.vis_y_bounds[bound_idx], plot.vis_y_bounds[bound_idx]))
+
+class ProbePlotMethodsTestCase(unittest.TestCase):
+    def setUp(self):
+
+        # generate working directory
+        self.test_dir = "test_dir"
+        if os.path.isdir(self.test_dir):
+            shutil.rmtree(self.test_dir)
+        os.mkdir(self.test_dir)
+
+        # generate input text files
+        solution_domain_setup(self.test_dir)
+
+        # generate initial condition file
+        self.sol_prim_in = np.array(
+            [
+                [1e6, 9e5],
+                [2.0, 1.0],
+                [1000.0, 1200.0],
+                [0.6, 0.4],
+            ]
+        )
+        np.save(os.path.join(self.test_dir, "test_init_file.npy"), self.sol_prim_in)
+
+        # set SystemSolver and SolutionDomain
+        self.solver = SystemSolver(self.test_dir)
+        self.sol_domain = SolutionDomain(self.solver)
+
+    def tearDown(self):
+
+        if os.path.isdir(self.test_dir):
+            shutil.rmtree(self.test_dir)
+
+    def test_probe_plot_plot(self):
+
+        sol_domain = self.sol_domain
+
+        # set plotting variables
+        vis_vars = [
+            "pressure",
+            "velocity",
+            "temperature",
+            "species-0",
+            "density",
+            "momentum",
+            "energy",
+            "density-species-0",
+        ]
+        vis_y_bounds = [
+            [0.8e6, 1.1e6],
+            [0.5, 2.5],
+            [900, 1300],
+            [0.3, 0.7],
+            [1.8, 2.6],
+            [1, 6],
+            [-2e7, -1.5e7],
+            [0.5, 1.75],
+        ]
+        vis_x_bounds = [[None, None]] * len(vis_vars)
+        
+        # overwrite SolutionDomain probe info
+        sol_domain.probe_vars = vis_vars
+        sol_domain.num_probe_vars = len(vis_vars)
+        self.solver.probe_vars = vis_vars
+
+        # loop probes
+        for probe_idx in range(sol_domain.num_probes):
+
+            start_idx = 0
+            # loop probe vars
+            for var_idx in range(len(vis_vars)):
+
+                # reset probe values
+                sol_domain.probe_vals = np.zeros((sol_domain.num_probes, sol_domain.num_probe_vars, self.solver.num_steps), dtype=REAL_TYPE)
+
+                if var_idx >= 9:
+                    start_idx += 1
+
+                plot = ProbePlot(
+                    self.solver.image_output_dir,
+                    0,
+                    self.solver.sim_type,
+                    sol_domain.probe_vars,
+                    vis_vars[start_idx:var_idx + 1],
+                    probe_idx,
+                    self.solver.num_probes,
+                    vis_x_bounds[start_idx:var_idx + 1],
+                    vis_y_bounds[start_idx:var_idx + 1],
+                    sol_domain.gas_model.species_names,
+                )
+
+                # update probes
+                for self.solver.iter in range(self.solver.num_steps):
+                    sol_domain.update_probes(self.solver)
+
[…]
