### 2021-11-30 8de8f6c
add test_snapshot_output
