### 2021-11-30 c0ca5a6
add stdout option to solver_params.inp to suppress iteration and residual print to STDOUT
