### 2021-11-30 d37d3ef
add driver test, will exclude shortly because this can hide coverage
