### 2021-11-30 14cbaed
move test_driver to a separate test, exclude driver from coverage
- since driver() touches everything, would hide coverage
- helps me to write more granular integration tests by targeting code untouched by driver test
- will still throw error in GHA if there is an issue in driver()
