+++ tests/integration_tests/test_suite.py
@@ -10,12 +10,11 @@ from test_reaction_model import test_finite_rate_irrev_reaction
 from test_flux.invisc_flux import test_invisc_flux
 from test_limiter import test_limiters
 from test_visualization import test_field_plot, test_probe_plot
-import test_driver
+
 
 loader = unittest.TestLoader()
 
-
-def integration_test_suite(output_mode=False):
+def integration_test_suite():
 
     suite = unittest.TestSuite()
 
@@ -41,12 +40,9 @@ def integration_test_suite(output_mode=False):
     # classes that depend on SolutionDomain
     suite.addTest(loader.loadTestsFromTestCase(test_limiters.LimiterMethodsTestCase))
     suite.addTest(loader.loadTestsFromTestCase(test_field_plot.FieldPlotInitTestCase))
-    # suite.addTest(loader.loadTestsFromTestCase(test_field_plot.FieldPlotMethodsTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_field_plot.FieldPlotMethodsTestCase))
     suite.addTest(loader.loadTestsFromTestCase(test_probe_plot.ProbePlotInitTestCase))
-    # suite.addTest(loader.loadTestsFromTestCase(test_probe_plot.ProbePlotMethodsTestCase))
-
-    # final driver test
-    suite.addTest(loader.loadTestsFromTestCase(test_driver.DriverTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_probe_plot.ProbePlotMethodsTestCase))
 
     return suite
 
@@ -84,4 +80,4 @@ if __name__ == "__main__":
         subprocess.call(os.path.join(localdir, "get_results.sh"))
 
     runner = unittest.TextTestRunner(verbosity=3)
-    runner.run(integration_test_suite(output_mode=output_mode))
+    runner.run(integration_test_suite())
