### 2021-12-01 138f643
add VisualizationGroup tests
