+++ tests/integration_tests/test_suite.py
@@ -9,7 +9,7 @@ from test_solution.test_solution_boundary import test_solution_boundary, test_so
 from test_reaction_model import test_finite_rate_irrev_reaction
 from test_flux.invisc_flux import test_invisc_flux
 from test_limiter import test_limiters
-from test_visualization import test_field_plot, test_probe_plot
+from test_visualization import test_field_plot, test_probe_plot, test_visualization_group
 
 
 loader = unittest.TestLoader()
@@ -43,6 +43,8 @@ def integration_test_suite():
     suite.addTest(loader.loadTestsFromTestCase(test_field_plot.FieldPlotMethodsTestCase))
     suite.addTest(loader.loadTestsFromTestCase(test_probe_plot.ProbePlotInitTestCase))
     suite.addTest(loader.loadTestsFromTestCase(test_probe_plot.ProbePlotMethodsTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_visualization_group.VisGroupInitTestCase))
+    suite.addTest(loader.loadTestsFromTestCase(test_visualization_group.VisGroupMethodsTestCase))
 
     return suite
 
