### 2021-12-01 86b59d3
tests were not returning correct exit code, hiding failed test
