+++ tests/integration_tests/test_solution/test_solution_domain.py
@@ -173,6 +173,10 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
                     np.load(os.path.join(self.output_dir, "sol_domain_subiter_sol_cons.npy")),
                 )
             )
+            print("self.sol_domain.sol_int.res")
+            print(self.sol_domain.sol_int.res)
+            print("From file")
+            print(np.load(os.path.join(self.output_dir, "sol_domain_subiter_res.npy")))
             self.assertTrue(
                 np.allclose(
                     self.sol_domain.sol_int.res, np.load(os.path.join(self.output_dir, "sol_domain_subiter_res.npy"))
@@ -191,6 +195,10 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
 
         else:
 
+            print("self.sol_domain.sol_int.sol_prim")
+            print(self.sol_domain.sol_int.sol_prim)
+            print("From file")
+            print(np.load(os.path.join(self.output_dir, "sol_domain_iter_sol_prim.npy")))
             self.assertTrue(
                 np.allclose(
                     self.sol_domain.sol_int.sol_prim,
