### 2021-12-01 2cd5800
temporarily remove offending SolutionDomain assertions
