### 2021-12-01 b9eb26a
remove offending driver assertions for now
