### 2021-12-01 a9fed73
add RomDomain initialization test
