### 2021-12-01 2c19009
cleanup some code debt
