+++ tests/unit_tests/test_misc_funcs.py
@@ -1,26 +1,24 @@
 import unittest
 import os
+import shutil
 
+from constants import TEST_DIR, del_test_dir
 import perform.misc_funcs as misc_funcs
 
 
 class MiscFuncsTestCase(unittest.TestCase):
     def setUp(self):
 
-        self.test_dir = "test_dir"
-
-        if os.path.isdir(self.test_dir):
-            os.rmdir(self.test_dir)
+        del_test_dir()
 
     def tearDown(self):
 
-        if os.path.isdir(self.test_dir):
-            os.rmdir(self.test_dir)
+        del_test_dir()
 
     def test_mkdir_shallow(self):
 
-        misc_funcs.mkdir_shallow("", self.test_dir)
-        self.assertTrue(os.path.isdir(self.test_dir))
+        misc_funcs.mkdir_shallow("", TEST_DIR)
+        self.assertTrue(os.path.isdir(TEST_DIR))
 
 
 if __name__ == "__main__":
