+++ tests/integration_tests/test_limiter/test_limiters.py
@@ -1,10 +1,9 @@
 import unittest
 import os
-import shutil
 
 import numpy as np
 
-from constants import solution_domain_setup
+from constants import TEST_DIR, del_test_dir, gen_test_dir, get_output_mode, solution_domain_setup
 from perform.system_solver import SystemSolver
 from perform.solution.solution_domain import SolutionDomain
 from perform.limiter.venkat_limiter import VenkatLimiter
@@ -14,31 +13,16 @@ from perform.limiter.barth_jesp_limiter import BarthJespCellLimiter, BarthJespFa
 class LimiterMethodsTestCase(unittest.TestCase):
     def setUp(self):
 
-        self.output_mode = bool(int(os.environ["PERFORM_TEST_OUTPUT_MODE"]))
-        self.output_dir = os.environ["PERFORM_TEST_OUTPUT_DIR"]
+        self.output_mode, self.output_dir = get_output_mode()
 
         # generate working directory
-        self.test_dir = "test_dir"
-        if os.path.isdir(self.test_dir):
-            shutil.rmtree(self.test_dir)
-        os.mkdir(self.test_dir)
+        gen_test_dir()
 
         # generate input text files
-        solution_domain_setup(self.test_dir)
-
-        # generate initial condition file
-        self.sol_prim_in = np.array(
-            [
-                [1e6, 9e5],
-                [2.0, 1.0],
-                [1000.0, 1200.0],
-                [0.6, 0.4],
-            ]
-        )
-        np.save(os.path.join(self.test_dir, "test_init_file.npy"), self.sol_prim_in)
+        solution_domain_setup()
 
         # set SystemSolver and SolutionDomain
-        self.solver = SystemSolver(self.test_dir)
+        self.solver = SystemSolver(TEST_DIR)
         self.sol_domain = SolutionDomain(self.solver)
 
         # calculate raw gradients
@@ -51,8 +35,7 @@ class LimiterMethodsTestCase(unittest.TestCase):
 
     def tearDown(self):
 
-        if os.path.isdir(self.test_dir):
-            shutil.rmtree(self.test_dir)
+        del_test_dir()
 
     def test_venkat_limiter(self):
 
