+++ tests/integration_tests/test_flux/invisc_flux/test_invisc_flux.py
@@ -3,7 +3,7 @@ import os
 
 import numpy as np
 
-from constants import CHEM_DICT_AIR
+from constants import get_output_mode, CHEM_DICT_AIR, SOL_PRIM_IN_AIR
 from perform.gas_model.calorically_perfect_gas import CaloricallyPerfectGas
 from perform.solution.solution_phys import SolutionPhys
 from perform.flux.invisc_flux.invisc_flux import InviscFlux
@@ -12,23 +12,13 @@ from perform.flux.invisc_flux.invisc_flux import InviscFlux
 class InviscFluxMethodsTestCase(unittest.TestCase):
     def setUp(self):
 
-        self.output_mode = bool(int(os.environ["PERFORM_TEST_OUTPUT_MODE"]))
-        self.output_dir = os.environ["PERFORM_TEST_OUTPUT_DIR"]
+        self.output_mode, self.output_dir = get_output_mode()
 
         self.chem_dict = CHEM_DICT_AIR
         self.gas = CaloricallyPerfectGas(self.chem_dict)
 
-        self.sol_prim_in = np.array(
-            [
-                [1e6, 1e5],
-                [2.0, 1.0],
-                [300.0, 400.0],
-                [0.4, 0.6],
-                [0.6, 0.4],
-            ]
-        )
         self.num_cells = 2
-        self.sol = SolutionPhys(self.gas, self.num_cells, self.sol_prim_in)
+        self.sol = SolutionPhys(self.gas, self.num_cells, SOL_PRIM_IN_AIR)
         self.flux_model = InviscFlux()
 
     def test_calc_inv_flux(self):

[…]
