### 2021-12-01 c297ae8
correct error in RomModel.sol_hist initialization, add prim var mapping testing
