### 2021-12-01 f95cd92
add cons var map tests
