+++ tests/integration_tests/test_rom/test_rom_variable_mapping.py
@@ -35,11 +35,11 @@ class RomPrimVarMappingMethodsTestCase(unittest.TestCase):
 
         # check get_variables_from_state
         sol_prim = self.var_mapping.get_variables_from_state(self.sol_domain)
-        self.assertTrue(np.array_equal(sol_prim, SOL_PRIM_IN_REACT))
+        self.assertTrue(np.allclose(sol_prim, SOL_PRIM_IN_REACT))
 
         # check get_variable_hist_from_state_hist
         sol_prim_hist = self.var_mapping.get_variable_hist_from_state_hist(self.sol_domain)
-        self.assertTrue(all([np.array_equal(sol_prim_arr, SOL_PRIM_IN_REACT) for sol_prim_arr in sol_prim_hist]))
+        self.assertTrue(all([np.allclose(sol_prim_arr, SOL_PRIM_IN_REACT) for sol_prim_arr in sol_prim_hist]))
 
     def test_prim_var_map_update_state(self):
 
@@ -59,7 +59,7 @@ class RomPrimVarMappingMethodsTestCase(unittest.TestCase):
         # check update_full_state
         self.var_mapping.update_full_state(self.sol_domain, self.rom_domain)
 
-        self.assertTrue(np.array_equal(self.sol_domain.sol_int.sol_prim, sol_prim_new))
+        self.assertTrue(np.allclose(self.sol_domain.sol_int.sol_prim, sol_prim_new))
 
         if self.output_mode:
             np.save(os.path.join(self.output_dir, "prim_var_map_update_full_sol_cons.npy"), self.sol_domain.sol_int.sol_cons)
@@ -72,9 +72,81 @@ class RomPrimVarMappingMethodsTestCase(unittest.TestCase):
         # check update_state_hist
         self.var_mapping.update_state_hist(self.sol_domain, self.rom_domain)
 
-        self.assertTrue(np.array_equal(self.sol_domain.sol_int.sol_hist_prim[0], sol_prim_new))
-        self.assertTrue(np.array_equal(self.sol_domain.sol_int.sol_hist_prim[1], SOL_PRIM_IN_REACT))
-        self.assertTrue(np.array_equal(self.sol_domain.sol_int.sol_hist_prim[2], SOL_PRIM_IN_REACT))
-        self.assertTrue(np.array_equal(self.sol_domain.sol_int.sol_hist_cons[0], self.sol_domain.sol_int.sol_cons))
-        self.assertTrue(np.array_equal(self.sol_domain.sol_int.sol_hist_cons[1], sol_cons_old))
-        self.assertTrue(np.array_equal(self.sol_domain.sol_int.sol_hist_cons[2], sol_cons_old))
+        self.assertTrue(np.allclose(self.sol_domain.sol_int.sol_hist_prim[0], sol_prim_new))
+        self.assertTrue(np.allclose(self.sol_domain.sol_int.sol_hist_prim[1], SOL_PRIM_IN_REACT))
+        self.assertTrue(np.allclose(self.sol_domain.sol_int.sol_hist_prim[2], SOL_PRIM_IN_REACT))
+        self.assertTrue(np.allclose(self.sol_domain.sol_int.sol_hist_cons[0], self.sol_domain.sol_int.sol_cons))
+        self.assertTrue(np.allclose(self.sol_domain.sol_int.sol_hist_cons[1], sol_cons_old))
+        self.assertTrue(np.allclose(self.sol_domain.sol_int.sol_hist_cons[2], sol_cons_old))
+
+
+class RomConsVarMappingMethodsTestCase(unittest.TestCase):
+    def setUp(self):
+
+        self.output_mode, self.output_dir = get_output_mode()
+
+        # generate working directory
+        gen_test_dir()
+
+        # generate input text files
+        solution_domain_setup()
[…]
