### 2021-12-01 21b857e
some minor testing additions
