+++ perform/misc_funcs.py
@@ -1,36 +1,6 @@
 """Miscellaneous functions that do not have a clear place in other modules."""
 
 import os
-import struct
-
-
-def write_to_file(fid, array, order="F"):
-    """Write NumPy arrays to binary file using struct.
-
-    I used to use this for comparisons against GEMS, but is not
-    terribly useful otherwise. Kept here just in case.
-
-    Args:
-        fid: File opened with open(), in mode 'wb'.
-        array: Array to be written to fid.
-        order: "C" or "F", order in which array will be flattened for output. "C" is row-major, "F" is column-major.
-    """
-
-    if array.ndim > 1:
-        array = array.flatten(order=order)
-    dtype = array.dtype
-    if dtype == "float64":
-        typeStr = "d"
-    elif dtype == "float32":
-        typeStr = "f"
-    elif dtype == "int32":
-        typeStr = "i"
-    elif dtype == "int16":
-        typeStr = "h"
-    else:
-        raise ValueError("Did not recognize array type " + dtype)
-
-    fid.write(struct.pack(typeStr * array.shape[0], *(array)))
 
 
 def mkdir_shallow(base_dir, new_dir_name):
