+++ tests/unit_tests/test_time_integrator/test_explicit_integrator.py
@@ -2,12 +2,8 @@ import unittest
 
 import numpy as np
 
-from perform.time_integrator.explicit_integrator import (
-    ExplicitIntegrator,
-    RKExplicit,
-    ClassicRK4,
-    SSPRK3,
-)
+from perform.time_integrator import get_time_integrator
+from perform.time_integrator.explicit_integrator import ExplicitIntegrator, RKExplicit
 
 
 class ExplicitTimeIntInitTestCase(unittest.TestCase):
@@ -54,7 +50,7 @@ class ClassicRK4InitTestCase(unittest.TestCase):
 
     def test_classic_rk4_init(self):
 
-        time_int = ClassicRK4(self.param_dict)
+        time_int = get_time_integrator("classic_rk4", self.param_dict)
 
         self.assertEqual(time_int.subiter_max, 4)
         self.assertEqual(len(time_int.rk_rhs), 4)
@@ -74,7 +70,7 @@ class ClassicRK4MethodsTestCase(unittest.TestCase):
         self.param_dict["time_order"] = 4
 
         # set up time integrator
-        self.time_int = ClassicRK4(self.param_dict)
+        self.time_int = get_time_integrator("classic_rk4", self.param_dict)
 
         self.rhs = np.array([[0, 3], [1, 4], [2, 5]])
         self.rk_rhs = [
@@ -122,7 +118,7 @@ class SSPRK3InitTestCase(unittest.TestCase):
 
     def test_ssp_rk3_init(self):
 
-        time_int = SSPRK3(self.param_dict)
+        time_int = get_time_integrator("ssp_rk3", self.param_dict)
 
         self.assertEqual(time_int.subiter_max, 3)
         self.assertEqual(len(time_int.rk_rhs), 3)
@@ -140,7 +136,7 @@ class SSPRK3MethodsTestCase(unittest.TestCase):
         self.param_dict["time_order"] = 3
 
         # set up time integrator
-        self.time_int = SSPRK3(self.param_dict)
+        self.time_int = get_time_integrator("ssp_rk3", self.param_dict)
 
         self.rhs = np.array([[0, 3], [1, 4], [2, 5]])
         self.rk_rhs = [
@@ -179,4 +175,72 @@ class SSPRK3MethodsTestCase(unittest.TestCase):
                 )
 
 
-# TODO: Jameson low storage
+class JamesonLowStoreInitTestCase(unittest.TestCase):
+    def setUp(self):
+        # set up param_dict
+        self.param_dict = {}
+        self.param_dict["dt"] = 1e-7
+        self.param_dict["time_scheme"] = "jameson_low_store"
+        self.param_dict["time_order"] = 4
+
+    def test_jameson_low_store_init(self):
+
+        time_int = get_time_integrator("jameson_low_store", self.param_dict)
+
+        self.assertEqual(time_int.subiter_max, 4)
+        self.assertEqual(len(time_int.rk_rhs), 4)
+        self.assertEqual(time_int.rk_a_vals.shape, (4, 4))
+        self.assertEqual(time_int.rk_b_vals.shape, (4,))
+        self.assertEqual(time_int.rk_c_vals.shape, (4,))
+
+
+class JamesonLowStoreMethodsTestCase(unittest.TestCase):
+    def setUp(self):
+        # set up param_dict
+        self.param_dict = {}
+        self.param_dict["dt"] = 1e-7
+        self.param_dict["time_scheme"] = "jameson_low_store"
+        self.param_dict["time_order"] = 4
+
+        # set up time integrator
[…]
