### 2021-12-02 15ca791
add uniform IC input test
