### 2021-12-02 8ad0725
add residual calculation tests
