### 2021-12-02 1fa80aa
add probe output tests, fix error in species probe writing
