### 2021-12-02 308fb65
add iter and final output tests
