+++ tests/integration_tests/test_solution/test_solution_domain.py
@@ -216,37 +216,7 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
 
     def test_probe_output(self):
 
-        sol_inlet = self.sol_domain.sol_inlet
-        sol_outlet = self.sol_domain.sol_outlet
-        sol_int = self.sol_domain.sol_int
-        probe_1_arr = np.array([
-            sol_inlet.sol_prim[0, 0],
-            sol_inlet.sol_prim[2, 0],
-            sol_inlet.sol_prim[3, 0],
-            sol_inlet.sol_cons[0, 0],
-        ])
-        probe_2_arr = np.array([
-            sol_int.sol_prim[0, 0],
-            sol_int.sol_prim[2, 0],
-            sol_int.sol_prim[3, 0],
-            sol_int.sol_cons[0, 0],
-        ])
-        probe_3_arr = np.array([
-            sol_outlet.sol_prim[0, 0],
-            sol_outlet.sol_prim[2, 0],
-            sol_outlet.sol_prim[3, 0],
-            sol_outlet.sol_cons[0, 0],
-        ])
-
-        probe_base = np.zeros((5, self.solver.num_steps), dtype=REAL_TYPE)
-        probe_base[0, :] = np.arange(1, self.solver.num_steps + 1) * self.solver.dt
-        probe_1_comp = probe_base.copy()
-        probe_2_comp = probe_base.copy()
-        probe_3_comp = probe_base.copy()
-        probe_1_comp[1:, :] = np.repeat(probe_1_arr[:, None], self.solver.num_steps, axis=-1)
-        probe_2_comp[1:, :] = np.repeat(probe_2_arr[:, None], self.solver.num_steps, axis=-1)
-        probe_3_comp[1:, :] = np.repeat(probe_3_arr[:, None], self.solver.num_steps, axis=-1)
-
+        probe_1_comp, probe_2_comp, probe_3_comp = gen_probe_comparisons(self.sol_domain, self.solver)
 
         for self.solver.iter in range(1, self.solver.num_steps + 1):
 
@@ -307,3 +277,167 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
         self.assertTrue(np.array_equal(probe_1, probe_1_comp))
         self.assertTrue(np.array_equal(probe_2, probe_2_comp))
         self.assertTrue(np.array_equal(probe_3, probe_3_comp))
+
+    def test_write_outputs(self):
+
+        sol_cons = self.sol_domain.sol_int.sol_cons
+        probe_1_comp, probe_2_comp, probe_3_comp = gen_probe_comparisons(self.sol_domain, self.solver)
+
+        self.solver.save_restarts = True
+        sol_int = self.sol_domain.sol_int
+
+        # testing write_iter_outputs()
+        for self.solver.iter in range(1, self.solver.num_steps + 1):
+            self.solver.time_iter += 1
+            self.solver.sol_time += self.solver.dt
+
+            self.sol_domain.write_iter_outputs(self.solver)
+
+            # check restart files
+            if ((self.solver.iter % self.solver.restart_interval) == 0):
+                restart_data = np.load(os.path.join(self.solver.restart_output_dir, "restart_file_" + str(self.solver.restart_iter - 1) + ".npz"))
+                self.assertTrue(np.array_equal(
+                    restart_data["sol_prim"],
+                    np.repeat(SOL_PRIM_IN_REACT[:, :, None], 2, axis=-1),    
+                ))
+                self.assertTrue(np.array_equal(
+                    restart_data["sol_cons"],
+                    np.repeat(sol_cons[:, :, None], 2, axis=-1),    
+                ))
[…]
