### 2021-12-02 47ddba9
correct error in not thresholding mass fractions on primitive mapping init
