### 2021-12-03 fdbbdc2
readd explicit linear Galerkin
