### 2021-12-03 b316369
add AutoencoderSpaceMapping, readd explicit nonlinear Galerkin
