+++ perform/rom/rom_space_mapping/linear_space_mapping.py
@@ -88,7 +88,7 @@ class LinearSpaceMapping(RomSpaceMapping):
         sol = np.reshape(sol, self.sol_shape, order="C")
         return sol
 
-    def calc_decoder_jacob_pinv(self):
+    def calc_decoder_jacob_pinv(self, code):
         """Returns pseudo-inverse of decoder Jacobian
 
         For linear, orthonormal trial basis, this is just the transpose of the trial basis
