+++ perform/rom/rom_variable_mapping/cons_variable_mapping.py
@@ -38,6 +38,11 @@ class ConsVariableMapping(RomVariableMapping):
 
         sol_domain.sol_int.update_state(from_prim=False)
 
+        # map back to account for any mass fraction thresholding
+        for rom_model in rom_domain.model_list:
+            var_idxs = rom_model.var_idxs
+            rom_model.sol[:, :] = sol_domain.sol_int.sol_cons[var_idxs, :].copy()
+
     def update_state_hist(self, sol_domain, rom_domain):
         """Collects model internal state history and updates relevant state history"""
 
@@ -53,3 +58,16 @@ class ConsVariableMapping(RomVariableMapping):
 
             # update primitive state history
             sol_int.sol_hist_prim[sol_idx] = sol_int.calc_prim_from_cons(sol_int.sol_hist_cons[sol_idx])
+
+            # threshold mass fractions
+            mass_fracs = sol_int.gas_model.get_mass_frac_array(sol_prim_in=sol_int.sol_hist_prim[sol_idx])
+            mass_fracs = sol_int.gas_model.calc_all_mass_fracs(mass_fracs, threshold=True)
+            sol_int.sol_hist_prim[sol_idx][3:, :] = mass_fracs[:-1, :]
+
+            # update conservative state history (again)
+            sol_int.sol_hist_cons[sol_idx] = sol_int.calc_cons_from_prim(sol_int.sol_hist_prim[sol_idx])
+
+            # map back to account for any mass fraction thresholding
+            for rom_model in rom_domain.model_list:
+                var_idxs = rom_model.var_idxs
+                rom_model.sol_hist[sol_idx][:, :] = sol_int.sol_hist_cons[sol_idx][var_idxs, :].copy()
