### 2021-12-05 6ff0e71
fix error in initializing ROM solution history
