### 2021-12-06 ecea9e1
readd nonlinear LSPG
