+++ perform/rom/rom_method/projection_method/mplsvt_projection.py
@@ -89,11 +89,14 @@ class MPLSVTProjection(ProjectionMethod):
             rhs: Right-hand side of low-dimensional linear solve.
         """
 
+        # compute (scaled) concatenated decoder Jacobians
+        decoder_jacob_concat, scaled_decoder_jacob_concat = self.assemble_concat_decoder_jacobs(sol_domain, rom_domain)
+
         # compute test basis
         if rom_domain.num_models == 1:
-            test_basis = np.array(res_jacob @ self.trial_basis_scaled_concat)
+            test_basis = np.array(res_jacob @ scaled_decoder_jacob_concat)
         else:
-            test_basis = (res_jacob @ self.trial_basis_scaled_concat).toarray()
+            test_basis = (res_jacob @ scaled_decoder_jacob_concat).toarray()
 
         # scaling
         res_scaled = np.zeros(test_basis.shape[0], dtype=REAL_TYPE)
