+++ perform/rom/rom_method/projection_method/projection_method.py
@@ -25,7 +25,7 @@ class ProjectionMethod(RomMethod):
         # Set up hyper-reduction, if necessary
         self.hyper_reduc = catch_input(rom_dict, "hyper_reduc", False)
         if self.hyper_reduc:
-            rom_domain.load_hyper_reduc(sol_domain)
+            rom_domain.load_hyper_reduc(rom_domain, sol_domain)
 
         # load and check gappy POD basis
         if self.hyper_reduc:
@@ -79,28 +79,6 @@ class ProjectionMethod(RomMethod):
         if (rom_domain.rom_dict["space_mapping"] == "linear") and (
             rom_domain.time_stepper.time_integrator.time_type == "implicit"
         ):
-            # if rom_domain.num_models == 1:
-            #     self.trial_basis_concat = rom_domain.model_list[0].space_mapping.trial_basis.copy()
-            #     self.trial_basis_scaled_concat = rom_domain.model_list[0].space_mapping.trial_basis_scaled.copy()
-            # else:
-            #     num_cells = sol_domain.mesh.num_cells
-            #     trial_basis_concat = np.zeros(
-            #         (num_cells * sol_domain.gas_model.num_eqs, rom_domain.latent_dim_total), dtype=REAL_TYPE
-            #     )
-            #     trial_basis_scaled_concat = trial_basis_concat.copy()
-            #     latent_dim_idx = 0
-            #     for model in rom_domain.model_list:
-            #         col_slice = np.s_[latent_dim_idx : latent_dim_idx + model.latent_dim]
-            #         for iter_idx, var_idx in enumerate(model.var_idxs):
-            #             row_slice_basis = np.s_[iter_idx * num_cells : (iter_idx + 1) * num_cells]
-            #             row_slice_concat = np.s_[var_idx * num_cells : (var_idx + 1) * num_cells]
-            #             trial_basis_concat[row_slice_concat, col_slice] = model.trial_basis[row_slice_basis, :]
-            #             trial_basis_scaled_concat[row_slice_concat, col_slice] = model.trial_basis_scaled[
-            #                 row_slice_basis, :
-            #             ]
-            #         latent_dim_idx += model.latent_dim
-            #     self.trial_basis_concat = np.array(trial_basis_concat)
-            #     self.trial_basis_scaled_concat = csr_matrix(trial_basis_scaled_concat)
 
             self.trial_basis_concat = self.concat_mapping(sol_domain, rom_domain, "trial_basis")
             self.trial_basis_scaled_concat = self.concat_mapping(
@@ -201,3 +179,269 @@ class ProjectionMethod(RomMethod):
             low_dim_vec = projector @ full_dim_vec
 
         return low_dim_vec
+
+    def load_hyper_reduc(self, rom_domain, sol_domain):
+        """Loads direct sampling indices and determines cell indices for hyper-reduction array slicing.
+
+        Numerous array slicing indices are required for various operations in efficiently computing
+        the non-linear RHS term, such as calculating fluxes, gradients, source terms, etc. as well as for computing
+        the RHS Jacobian if required. These slicing arrays are first generated here based on the initial sampling
+        indices, but may later be updated during sampling adaptation.
+
+        Todos:
+            Many of these operations should be moved to their own separate functions when
+            recomputing sampling for adaptive sampling.
+
+        Args:
[…]
