+++ perform/rom/rom_domain.py
@@ -1,11 +1,7 @@
 import os
 from time import sleep
 
-import numpy as np
-
-from perform.constants import REAL_TYPE
 from perform.input_funcs import catch_list, catch_input
-from perform.solution.solution_phys import SolutionPhys
 from perform.rom import get_ml_library, get_rom_method, get_variable_mapping, get_time_stepper
 from perform.rom.rom_model import RomModel
 
@@ -176,262 +172,3 @@ class RomDomain:
 
             model.code_hist[1:] = model.code_hist[:-1]
             model.code_hist[0] = model.code.copy()
-
-    def load_hyper_reduc(self, sol_domain):
-        """Loads direct sampling indices and determines cell indices for hyper-reduction array slicing.
-
-        Numerous array slicing indices are required for various operations in efficiently computing
-        the non-linear RHS term, such as calculating fluxes, gradients, source terms, etc. as well as for computing
-        the RHS Jacobian if required. These slicing arrays are first generated here based on the initial sampling
-        indices, but may later be updated during sampling adaptation.
-
-        Todos:
-            Many of these operations should be moved to their own separate functions when
-            recomputing sampling for adaptive sampling.
-
-        Args:
-            sol_domain: SolutionDomain with which this RomDomain is associated.
-        """
-
-        # TODO: add some explanations for what each index array accomplishes
-
-        # load and check sample points
-        samp_file = catch_input(self.rom_dict, "samp_file", "")
-        assert samp_file != "", "Must supply samp_file if performing hyper-reduction"
-        samp_file = os.path.join(self.model_dir, samp_file)
-        assert os.path.isfile(samp_file), "Could not find samp_file at " + samp_file
-
-        # Indices of directly sampled cells, within sol_prim/cons
-        # NOTE: assumed that sample indices are zero-indexed
-        sol_domain.direct_samp_idxs = np.load(samp_file).flatten()
-        sol_domain.direct_samp_idxs = (np.sort(sol_domain.direct_samp_idxs)).astype(np.int32)
-        sol_domain.num_samp_cells = len(sol_domain.direct_samp_idxs)
-        assert (
-            sol_domain.num_samp_cells <= sol_domain.mesh.num_cells
-        ), "Cannot supply more sampling points than cells in domain."
-        assert np.amin(sol_domain.direct_samp_idxs) >= 0, "Sampling indices must be non-negative integers"
-        assert (
-            np.amax(sol_domain.direct_samp_idxs) < sol_domain.mesh.num_cells
-        ), "Sampling indices must be less than the number of cells in the domain"
-        assert (
-            len(np.unique(sol_domain.direct_samp_idxs)) == sol_domain.num_samp_cells
-        ), "Sampling indices must be unique"
-
-        # Compute indices for inviscid flux calculations
-        # NOTE: have to account for fact that boundary cells are prepended/appended
-        # Indices of "left" cells for flux calcs, within sol_prim/cons_full
-        sol_domain.flux_samp_left_idxs = np.zeros(2 * sol_domain.num_samp_cells, dtype=np.int32)
-        sol_domain.flux_samp_left_idxs[0::2] = sol_domain.direct_samp_idxs
-        sol_domain.flux_samp_left_idxs[1::2] = sol_domain.direct_samp_idxs + 1
-
-        # Indices of "right" cells for flux calcs, within sol_prim/cons_full
[…]
