### 2021-12-06 c2d52b6
add LinearSpaceMapping tests
