+++ tests/integration_tests/test_solution/test_solution_domain.py
@@ -123,7 +123,9 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
             np.save(os.path.join(self.output_dir, "sol_domain_res_jacob_phys_time.npy"), res_jacob)
 
         else:
-            self.assertTrue(np.allclose(res_jacob, np.load(os.path.join(self.output_dir, "sol_domain_res_jacob_phys_time.npy"))))
+            self.assertTrue(
+                np.allclose(res_jacob, np.load(os.path.join(self.output_dir, "sol_domain_res_jacob_phys_time.npy")))
+            )
 
     def test_advance_subiter(self):
 
@@ -209,10 +211,9 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
         # go through get_initial_conditions to catch some more lines
         sol_prim_init = get_initial_conditions(self.sol_domain, self.solver)
 
-        self.assertTrue(np.array_equal(
-            sol_prim_init[:, :, 0],
-            np.array([[1e4, 2e4], [0.5, 10.0], [5000.0, 400.0], [0.6, 0.2]])
-        ))
+        self.assertTrue(
+            np.array_equal(sol_prim_init[:, :, 0], np.array([[1e4, 2e4], [0.5, 10.0], [5000.0, 400.0], [0.6, 0.2]]))
+        )
 
     def test_probe_output(self):
 
@@ -228,52 +229,106 @@ class SolutionDomainMethodsTestCase(unittest.TestCase):
                 self.solver.iter != self.solver.num_steps
             ):
                 self.sol_domain.write_probes(self.solver, intermediate=True, failed=False)
-                
-                probe_1_itmdt = np.load(os.path.join(self.solver.probe_output_dir, "probe_pressure_temperature_species-0_density_1_" + self.solver.sim_type + "_ITMDT.npy"))
-                probe_2_itmdt = np.load(os.path.join(self.solver.probe_output_dir, "probe_pressure_temperature_species-0_density_2_" + self.solver.sim_type + "_ITMDT.npy"))
-                probe_3_itmdt = np.load(os.path.join(self.solver.probe_output_dir, "probe_pressure_temperature_species-0_density_3_" + self.solver.sim_type + "_ITMDT.npy"))
-                
-                self.assertTrue(np.array_equal(probe_1_itmdt, probe_1_comp[:, :self.solver.iter]))
-                self.assertTrue(np.array_equal(probe_2_itmdt, probe_2_comp[:, :self.solver.iter]))
-                self.assertTrue(np.array_equal(probe_3_itmdt, probe_3_comp[:, :self.solver.iter]))
-                
+
+                probe_1_itmdt = np.load(
+                    os.path.join(
+                        self.solver.probe_output_dir,
+                        "probe_pressure_temperature_species-0_density_1_" + self.solver.sim_type + "_ITMDT.npy",
+                    )
+                )
+                probe_2_itmdt = np.load(
+                    os.path.join(
+                        self.solver.probe_output_dir,
+                        "probe_pressure_temperature_species-0_density_2_" + self.solver.sim_type + "_ITMDT.npy",
+                    )
+                )
+                probe_3_itmdt = np.load(
+                    os.path.join(
+                        self.solver.probe_output_dir,
+                        "probe_pressure_temperature_species-0_density_3_" + self.solver.sim_type + "_ITMDT.npy",
+                    )
+                )
+
+                self.assertTrue(np.array_equal(probe_1_itmdt, probe_1_comp[:, : self.solver.iter]))
+                self.assertTrue(np.array_equal(probe_2_itmdt, probe_2_comp[:, : self.solver.iter]))
[…]
