+++ .github/workflows/perform_basic.yml
@@ -28,6 +28,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -e .
+          pip install tensorflow==2.7.0
           pip install coverage coveralls
 
       - name: Run unit tests

[…]
