### 2021-12-06 bb2ea5b
add AutoencoderSpaceMapping tests
