### 2021-12-06 103fd44
add GalerkinProjection tests
