### 2021-12-06 58e851d
exclude residual comparison for now
