### 2021-12-06 6f85bd8
add test running script, clean things up
