### 2021-12-06 7759337
add coverage and DOI badges
