+++ paper/paper.bib
@@ -121,11 +121,20 @@
 	Year = {2017}
 }
 
-@article{Huang2020,
+@article{Huang2021,
 	author = {Cheng Huang and Christopher R. Wentland and Karthik Duraisamy and Charles Merkle},
-	title = {Model reduction for multi-scale transport problems using structure-preserving least-squares projections with variable transformation},
-	year = {2020},
-	journal = {arXiv:2011.02072v3}
+	title = {Model reduction for multi-scale transport problems using model-form preserving least-squares projections with variable transformation},
+	year = {2021},
+	journal = {Journal of Computational Physics},
+	volume = {448},
+	doi = {10.1016/j.jcp.2021.110742}
+}
+
+@article{Rezaian2021,
+	author = {Elnaz Rezaian and Cheng Huang and Karthik Duraisamy},
+	title = {Non-intrusive Balancing Transformation of Highly Stiff Systems with Lightly-damped Impulse Response},
+	year = {2021},
+	journal = {arXiv:2109.10408}
 }
 
 @article{Everson1995,
