+++ paper/paper.md
@@ -13,7 +13,7 @@ authors:
 affiliations:
  - name: Department of Aerospace Engineering, University of Michigan
    index: 1
-date: March 25, 2021
+date: December 6, 2021
 bibliography: paper.bib
 ---
 
@@ -27,9 +27,7 @@ The Prototyping Environment for Reacting Flow Order Reduction Methods (PERFORM) 
 
 The ROM community spans many scientific disciplines, and efforts to build and release tools for ROM research and applications have been commensurately broad. Some open-source projects provide implementations of standard ROM methods within a specific flow solver [@ithacasem;@Stabile2017CAIM;@Stabile2017CAF;@Hesthaven2016;@acceleratecfd]. Other libraries have been developed to provide generic interfaces with dynamical system solvers and supply ROM capabilities, most notably Pressio [@Rizzi2020] and pyMOR [@Milk2016]. These libraries allow CFD practitioners to test a suite of standard ROM methods with an existing CFD solver. However, these libraries may not be useful to ROM method developers, who may find it difficult and prohibitively time-consuming to implement new ROM methods in a very complex software environment or in a low-level programming language. Further, those who wish to test ROM methods on combustion problems may not have access to a robust and efficient combustion CFD solver (many are closed-source) or the know-how to use general open-source solvers such as OpenFOAM. 
 
-PERFORM aims to ease some of these difficulties by packaging a one-dimensional compressible reacting flow solver with a flexible object-oriented ROM framework. It trades flexibility of application for flexibility of ROM development, focusing on a single class of problems while making the code extremely accessible to ROM researchers of all backgrounds. For one, PERFORM is written purely in Python, a widely-used, easily-learned, and flexible high-level programming language. Additionally, ROMs in PERFORM are structured within a class hierarchy which maximizes code re-use and minimizes developer effort in implementing variations of ROM methods. The class hierarchy for projection-based ROMs shown below displays this modular framework, where the lowest level classes (e.g. `LinearGalerkinProj`, `LinearLSPGProj`, `LinearSPLSVTProj`) implement two to three instance methods which differ by only a few lines of code. 
-
-![Projection ROM class hierarchy](./proj_rom_class_hierarchy.jpg)
+PERFORM aims to ease some of these difficulties by packaging a one-dimensional compressible reacting flow solver with a flexible object-oriented ROM framework. It trades flexibility of application for flexibility of ROM development, focusing on a single class of problems while making the code extremely accessible to ROM researchers of all backgrounds. For one, PERFORM is written purely in Python, a widely-used, easily-learned, and flexible high-level programming language. Additionally, ROMs in PERFORM are structured within a class hierarchy which maximizes code re-use and minimizes developer effort in implementing variations of ROM methods.
 
 In most cases, only a basic understanding of Python classes and NumPy/SciPy operations is required to implement new ROM methods. Further, all elements of the underlying reacting flow solver are separated from the ROM classes, ensuring that ROM method developers do not need to interact with the solver routines. As such, the learning curve for understanding PERFORM's ROM mechanics and implementing new ROM methods is relatively gentle. Note that although pyMOR [@Milk2016] similarly provides a pure Python ROM framework, it aims to be far more generalizable to any class of problems, be compatible with a host of low-level computational and data handling libraries, and enable parallel computing. While this makes pyMOR much more flexible and efficient, it can also make ROM prototyping and code maintenance much more cumbersome. 
 
@@ -37,7 +35,7 @@ PERFORM's public repository comes with several benchmark cases which are ready t
 
 # Features
 
[…]
