### 2021-12-06 #3428 @cwentland0
Hi all, just wanted to give an update. I have put together a fairly thorough testing framework of unit, integration, and regression tests which can be found in the `tests/` directory. Instructions for manually running the tests have been included in the `README`. Additionally, I have added continuous integration via GitHub Actions, which automatically executes unit and integration tests and reports code coverage via the Coveralls API. Regression tests are run nightly and aren't included in the code coverage. Hopefully this covers the testing requirement.

Additionally, I have made minor changes to the paper to a) sanitize it of any specific mention of class names, since those might change and evolve over time, and b) add a citation for a recent paper that has used PERFORM. No other substantive changes were made to the paper. I had to make a Zenodo DOI for the repository to make the code citeable for collaborators, I think I recall that this can be substituted for the JOSS paper DOI in the event of acceptance?

Thank you for your patience!
