### 2021-12-06 0d24591
fix tensorflow failing FOM integration tests, add warning in README
