### 2021-12-07 e8f19ec
convert solver theory document to LaTeX, remove PDF
