+++ doc/solver_theory/physics.tex
@@ -1,0 +1,241 @@
+\newpage
+
+\section{Physics}\label{sec:physics}
+
+\subsection{Governing Equations}\label{sec:govEqs}
+The governing equations for the 1D Navier-Stokes equations with species transport and reactions are
+\be\label{eq:govEqs}
+    \pde{\consVec}{t} + \pde{\fluxVec}{x} - \pde{\fluxVec_v}{x} = \sourceVec.
+\ee
+The conservative state \consVec, the inviscid flux vector \fluxVec, the viscous flux vector $\fluxVec_v$, and the source term \sourceVec\ are given by
+\begin{equation}\label{eq:govEqsTerms}
+    \consVec = 
+    \begin{bmatrix}
+        \rho \\ \rho u \\ \rho h^0 - p \\ \rho Y_\speciesIdx
+    \end{bmatrix}, \quad
+    \fluxVec = 
+    \begin{bmatrix}
+        \rho u \\  \rho u^2 + p \\ \rho h^0 u \\ \rho Y_\speciesIdx 
+    \end{bmatrix}, \quad
+    \fluxVec_v = 
+    \begin{bmatrix}
+        0 \\ \tau \\ u \tau - q \\ -\rho V_\speciesIdx Y_\speciesIdx
+    \end{bmatrix}, \quad
+    \mathbf{s} = 
+    \begin{bmatrix}
+        0 \\ 0 \\ 0 \\ \dot{\omega}_\speciesIdx
+    \end{bmatrix}
+\end{equation}
+where $\rho$ is density, $u$ is velocity, $h^0$ is stagnation enthalpy, $p$ is static pressure, and $Y_\speciesIdx$ is the mass fraction of the \speciesIdx th chemical species. For a system with \numSpecies\ chemical species, only $\numSpecies - 1$ species transport equations are solved, as the final species mass fraction $Y_\numSpecies$ can be computed from the fact that all mass fractions must sum to unity,
+\be\label{eq:massFracSumOne}
+    Y_\numSpecies = 1 - \sum_{\speciesIdx=1}^{\numSpecies-1} Y_\speciesIdx
+\ee
+The stagnation enthalpy $h^0$ is given by
+\begin{equation}
+    h^0 = \sum_{\speciesIdx = 1}^{\numSpecies} h_\speciesIdx Y_\speciesIdx + \frac{1}{2} u^2,
+\end{equation}
+where $h_\speciesIdx$ is the enthalpy of the \speciesIdx th species. The shear stress $\tau$ is given by
+\be
+   \tau = \frac{4}{3} \mu \pde{u}{x},
+\ee
+where $\mu$ is the mixture dynamic viscosity. The heat flux $q$ is given by
+\be
+    q = -\thermCond \pde{T}{x} + \rho \sum_{\speciesIdx}^{\numSpecies} V_\speciesIdx Y_\speciesIdx h_\speciesIdx,
+\ee
+where $\thermCond$ is the mixture thermal conductivity. The diffusion velocity term $V_\speciesIdx Y_\speciesIdx$, which also appears in the species transport viscous flux, is approximated as
+\be
+    V_\speciesIdx Y_\speciesIdx = -D_{\speciesIdx, M} \pde{Y_\speciesIdx}{x}
+\ee
+where $D_{\speciesIdx, M}$ is the mass diffusion coefficient for the \speciesIdx th species diffusing into the mixture. 
+
[…]
