+++ doc/solver_theory/roms.tex
@@ -1,0 +1,4 @@
+\newpage
+\section{Reduced-Order Modeling}\label{sec:roms}
+
+This section will be updated shortly, check back soon!
