+++ doc/solver_theory/jacobians.tex
@@ -1,0 +1,4 @@
+\newpage
+\section{Jacobians}\label{sec:jacobIndex}
+
+This section will be updated shortly, check back soon!
