+++ doc/solver_theory/numerics.tex
@@ -1,0 +1,520 @@
+\newpage
+\section{Numerics}
+
+In this section, we describe the various methods used to discretize the governing PDE in both time and space. Details of new methods will be added when they are implemented in \verb|PERFORM|.
+
+\subsection{Spatial Discretization}
+
+A cell-centered finite volume formulation is used to discretize the one-dimensional spatial domain into \numCells\ discrete control volumes (or ``cells''). The spatial coordinate of the center of the \coordIdx th cell is denoted by \cellCentX, and the coordinate of the cell's left ``face'' and right ``face'' are given by \faceXLeft\ and \faceXRight, respectively. The general form of the finite volume formulation is given by 
+\be
+    \ode{}{\timeVar} \int_{\Omega} \consVec \; \dVol + \int_{\partial \Omega} \nabla \cdot (\fluxVec - \fluxVec_v) \cdot \vec{n} \; \text{ds} = \int_{\Omega} \sourceVec \; \dVol
+\ee
+where $\Omega$ is the interior domain of the finite volume cell, and $\partial \Omega$ is the surface of the cell. Simplifying this formulation to one spatial dimension results in
+\be
+ \ode{}{\timeVar} \int_{\faceXLeft}^{\faceXRight} \consVec \; \dx + \left( (\fluxVec - \fluxVec_v)_{\faceXRight} - (\fluxVec - \fluxVec_v)_{\faceXLeft} \right) = \int_{\faceXLeft}^{\faceXRight} \sourceVec \; \dx
+\ee
+where, for example, the term $\fluxVec_{\faceXLeft}$ is the inviscid flux through the \coordIdx th cell's left face. The calculation of these fluxes is discussed in Section~\ref{sec:fluxSchemes}. 
+
+Evaluating the volume integrals,
+\be
+    \Dx_\coordIdx \ode{\consVec_\coordIdx}{\timeVar} + \left( (\fluxVec - \fluxVec_v)_{\faceXRight} - (\fluxVec - \fluxVec_v)_{\faceXLeft} \right) = \Dx_\coordIdx \; \sourceVec_\coordIdx
+\ee
+where the cell length is $\Dx_\coordIdx = \faceXRight - \faceXLeft$. We can rearrange terms to arrive at
+\be
+    \ode{\consVec_\coordIdx}{\timeVar} = \frac{1}{\Dx_\coordIdx} \left( (\fluxVec - \fluxVec_v)_{\faceXLeft} - (\fluxVec - \fluxVec_v)_{\faceXRight} \right) + \sourceVec_\coordIdx
+\ee
+With this definition, we can extend the definition of the conservative solution, flux, and source vectors to include those at every cell in the spatial domain. For example, the conservative state would be defined as
+\be
+    \consVec := [\rho_1, \hdots, \rho_{\numCells}, (\rho u)_1, \hdots, (\rho u)_{\numCells}, (\rho h^0 - p)_1, \hdots, (\rho h^0 - p)_{\numCells}, (\rho Y_\speciesIdx)_1, \hdots, (\rho Y_\speciesIdx)_{\numCells}]^\top
+\ee
+and the flux/source terms are defined accordingly. Thus, the ODE governing the entire discrete system state is given by
+\be
+    \ode{\consVec}{\timeVar} = \frac{1}{\Delta\text{\textbf{x}}} \left( (\fluxVec - \fluxVec_v)_{L} - (\fluxVec - \fluxVec_v)_{R} \right) + \sourceVec
+\ee
[…]
