### 2021-12-08 3f0a48b
update most FOM text descriptions and input file descriptions
