+++ doc/index.rst
@@ -1,8 +1,8 @@
 PERFORM
 ===========
-**PERFORM** is a combination 1D compressible reacting flow solver and modular reduced-order model (ROM) framework, designed to provide a simple, easy-to-use testbed for members of the ROM community to quickly prototype and test new methods on challenging (yet computationally-manageable) reacting flow problems. The baseline solver is based off the **General Equations and Mesh Solver (GEMS)**:cite:p:`Ding2004`, a legacy Fortran 3D reacting flow solver originally developed by Li Ding and Guoping Xia at Purdue University. This code has obviously been simplified to one-dimensional flows, but it aims to be a) open-source, freely available to copy and modify for everyone, and b) much more approachable for researchers outside the high-performance computing community.  We hope that this tool lowers the barrier to entry for researchers from a variety of fields to develop novel ROM methods and benchmark them against an interesting set of reacting flow configurations.
+**PERFORM** is a combination 1D compressible reacting flow solver and modular reduced-order model (ROM) framework, designed to provide a simple, easy-to-use testbed for members of the ROM community to quickly prototype and test new methods on challenging (yet computationally-manageable) reacting flow problems. The baseline solver is based off the **General Equations and Mesh Solver (GEMS)**:cite:p:`Ding2004`, a Fortran 3D reacting flow solver originally developed by Li Ding and Guoping Xia at Purdue University. This code has obviously been simplified to one-dimensional flows, but it aims to be a) open-source, freely available to copy and modify for everyone, and b) much more approachable for researchers outside the high-performance computing community.  We hope that this tool lowers the barrier to entry for researchers from a variety of fields to develop novel ROM methods and benchmark them against an interesting set of reacting flow configurations.
 
-This documentation serves as a reference for users to get up and running with **PERFORM**, understand the underlying solver (albeit very superficially), and become acquainted with implementing new ROM routines and evaluating their performance. The mathematical theory behind the solver and ROM methods is only touched on briefly; we refer the interested reader to the PDF document supplied in ``perform/doc/PERFORM_solver_doc.pdf`` for more details. This website and the theory documentation is very much a work-in-progress and will be updated regularly.
+This documentation serves as a reference for users to get up and running with **PERFORM**, understand the underlying solver (albeit very superficially), and become acquainted with implementing new ROM routines and evaluating their performance. The mathematical theory behind the solver and ROM methods is only touched on briefly; we refer the interested reader to compile the solver documentation from ``perform/doc/solver_theory/main.tex`` for more details. This website and the theory documentation is very much a work-in-progress and will be updated regularly.
 
 Data-driven Modeling for Complex Fluid Physics
 ---------------------------------------------------------
@@ -48,8 +48,6 @@ Christopher R. Wentland acknowledges support from the US Air Force Office of Sci
 	roms/rominputs
 	roms/linearsub
 	roms/nonlinsub
-	roms/liftlearn
-	roms/caetcn
 
 .. toctree::
 	:maxdepth: 3
