+++ doc/userguide/paramindex.rst
@@ -8,13 +8,13 @@ solver_params.inp
 -----------------
 See :ref:`solverparams-label` for variable types, default values, and units (where applicable).
 
-* ``mesh_file``: Absolute path to mesh file.
+* ``mesh_file``: Path to mesh file. Permits absolute path, or relative path from working directory.
 
-* ``chem_file``: Absolute path to chemistry file.
+* ``chem_file``: Path to chemistry file. Permits absolute path, or relative path from working directory.
 
-* ``init_file``: Absolute path to full initial primitive state profile (stored in a ``*.npy`` NumPy binary file) to initialize the unsteady solution from. If ``init_from_restart = True``, this parameter will be ignored and the unsteady solution will be initialized from a restart file.
+* ``init_file``: Path to full initial primitive state profile (stored in a ``*.npy`` NumPy binary file) to initialize the unsteady solution from. Permits absolute path, or relative path from working directory. If ``ic_params_file`` is set, or ``init_from_restart = True``, this parameter will be ignored.
 
-* ``ic_params_file``: Absolute path to left/right (step function) primitive state parameters file to initialize the unsteady solution from. If ``init_file`` is set or ``init_from_restart = True``, this parameter will be ignored.
+* ``ic_params_file``: Path to left/right (step function) primitive state parameters file to initialize the unsteady solution from. Permits absolute path, or relative path from working directory. If ``init_from_restart = True``, this parameter will be ignored.
 
 * ``dt``: Fixed time step size for numerical time integration.
 
@@ -56,7 +56,7 @@ See :ref:`solverparams-label` for variable types, default values, and units (whe
 
 * ``grad_limiter``: Name of the gradient limiter to use when computing higher-order face reconstructions. Please see the solver documentation for details on each scheme.
 
-  * Valid options: ``barth``, ``venkat``
+  * Valid options: ``barth_cell``, ``barth_face``, ``venkat``
 
 * ``bound_cond_inlet``: Name of the boundary condition to apply at the inlet. For details on each boundary condition, see the solver documentation. For required input parameters for a given boundary condition, see :ref:`inletbcs-label`.
 
@@ -100,6 +100,8 @@ See :ref:`solverparams-label` for variable types, default values, and units (whe
 
 * ``vel_add``: Velocity to be added to the entire initial condition velocity field. Accepts negative values.
 
+* ``stdout``: Boolean flag to specify whether to print iteration counts and residual norms to STDOUT.
+
 * ``res_norm_prim``: List of values by which to normalize each field of the :math:`\ell^2` and :math:`\ell^1` residual norms before averaging across all fields. They are order by pressure, velocity, temperature, and then all species mass fractions except the last. This ensures that the norms of each residual field contribute roughly equally to the average norm used to determine Newton's method convergence.
 
 * ``source_off``: Boolean flag to specify whether to apply the reaction source term. This is ``False`` by default; setting it manually to ``True`` turns off the source term. This can save computational cost for non-reactive cases.
@@ -117,7 +119,7 @@ See :ref:`solverparams-label` for variable types, default values, and units (whe
 * ``probe_vars``: A list of fields to be probed at each specified probe location.
 
   * Valid for all probes: ``"pressure"``, ``"velocity"``, ``"temperature"``, ``"density"``, ``"momentum"``, ``"energy"``, ``"species_X"``, ``"density-species_X"`` (where ``X`` is replaced by the integer number of the desired chemical species to be probed, e.g. ``"species_2"`` for the second species specified in the chemistry file).
[…]
