+++ doc/userguide/contributions.rst
@@ -3,4 +3,4 @@ Issues and Contributing
 
 If you experience errors or unexpected solver behavior when running **PERFORM**, please first double-check your input parameters and use this documentation as a reference for proper input file formatting. If problems persist, please create a new issue on the GitHub repository, and we'll do our best to resolve it. However, if a submitted issue is related to a significant *expansion* of code capabilities (e.g. adding a new ROM model or flux scheme), we probably won't work on it. 
 
-On the other hand, if you would like to personally work to expand the code and contribute to **PERFORM**, first of all thank you! Please fork the main repository under your own GitHub account, create a new branch on your fork, and implement your contributions. When you're finished, create a pull request against the main repository and your changes will be reviewed before being merged into the main repository. Automatic testing has not yet been implemented for the repository, so in the meantime, make sure that your changes are compatible with *all* of the example cases before creating a pull request.
+On the other hand, if you would like to personally work to expand the code and contribute to **PERFORM**, first of all thank you! Please fork the repository under your own GitHub account and implement your contributions. When you're finished, create a pull request against the main repository and your changes will be reviewed before being merged. You can manually check beforehand that your changes do not break code by running Bash script ``perform/tests/run_tests.sh``, which will execute unit, integration, and regression tests. You can also let GitHub Actions automatically run these tests when you push changes to the remote repo or submit a pull request.
