### 2021-12-27 4318802
correct some errors in reaction model theory docs
