+++ doc/solver_theory/physics.tex
@@ -189,19 +189,17 @@ where $\text{Sc}_\speciesIdx$ is the Schmidt number of the \speciesIdx th specie
 \newpage
 \subsection{Reaction Models}\label{sec:reacModels}
 
-Various models exist for computing the species production rate $\dot{\omega}_\speciesIdx$. Of course, the most accurate method would be elementary reaction mechanisms, but these often involve thousands of species and reactions, pressure corrections and are thus infeasible even for 1D problems. Reduced mechanisms and global reactions offer two comparatively inexpensive alternatives. We describe both in turn.
+Various models exist for computing the species production rate $\dot{\omega}_\speciesIdx$. We describe available reaction models below.
 
-\subsubsection{Reduced Mechanisms}
+\subsubsection{Finite Rate Mechanisms}
 
-This section describes the modeling of reaction source terms by reduced mechanisms. The same methodology may be applied to elementary reaction mechanisms, but we assume that some empirical, reduced mechanism is used to grossly reduce the number of modeled species and reactions to $\MC{O}(10-100)$.
-
-We begin by supposing that a reaction mechanism is made up of \numReacs\ reactions governing the production and consumption of \numSpecies\ chemical species. The \reacIdx th reaction of this mechanism can be described by the formula
+This section describes the modeling of reaction source terms by finite rate mechanisms. We begin by supposing that a reaction mechanism is made up of \numReacs\ reactions governing the production and consumption of \numSpecies\ chemical species. The \reacIdx th reaction of this mechanism can be described by the formula
 \be
-   \sum_{\speciesIdx=1}^{\numSpecies} \nu_{\speciesIdx,\reacIdx}' \chi_\speciesIdx \xrightleftharpoons[k_{f,\reacIdx}]{k_{r,\reacIdx}} \sum_{\speciesIdx=1}^{\numSpecies} \nu_{\speciesIdx,\reacIdx}'' \chi_\speciesIdx
+   \sum_{\speciesIdx=1}^{\numSpecies} \nu_{\speciesIdx,\reacIdx}' \chi_\speciesIdx \xrightleftharpoons[k_{F,\reacIdx}]{k_{R,\reacIdx}} \sum_{\speciesIdx=1}^{\numSpecies} \nu_{\speciesIdx,\reacIdx}'' \chi_\speciesIdx
 \ee
-where $\chi_\speciesIdx$ is the chemical formula for the \speciesIdx th chemical species, and $\nu_{\speciesIdx,\reacIdx}'$ and $\nu_{\speciesIdx,\reacIdx}''$ are the stoichiometric coefficients of the reactants and products, respectively, for the \speciesIdx th species in the \reacIdx th reaction. The coefficients $k_{f,\reacIdx}$ and $k_{r,\reacIdx}$ forward and reverse reaction rates for the \reacIdx th reaction. For example, the single-step combustion of methane can be described by
+where $\chi_\speciesIdx$ is the chemical formula for the \speciesIdx th chemical species, and $\nu_{\speciesIdx,\reacIdx}'$ and $\nu_{\speciesIdx,\reacIdx}''$ are the stoichiometric coefficients of the reactants and products, respectively, for the \speciesIdx th species in the \reacIdx th reaction. The coefficients $k_{F,\reacIdx}$ and $k_{R,\reacIdx}$ are the forward and reverse reaction rates for the \reacIdx th reaction, respectively. For example, the single-step combustion of methane can be described by
[…]
