### 2021-12-27 fb83cfa
a couple minor corrections to RTD FOM pages
