### 2022-01-21 #3428 @kyleniemeyer
@cwentland0 thanks for the update! (And sorry it took so long for me to get back to this.) Yes, we will use the Zenodo DOI of the repo, though you will also get one for the paper itself.

@Himscipy can you possibly take another look at the updates, to see if your issues are now satisfied?

@clemaitre58 are you still able to review this submission?
