### 2022-01-25 d60a2ce
probes include initial conditions, snapshots save non-solutions variables correctly
