### 2022-02-01 6b819ad
minor change to solver theory intro
