+++ doc/solver_theory/introduction.tex
@@ -5,7 +5,7 @@ This code was originally written to be a sort of low-fidelity 1D Python port of 
 
 Although this port is still ``low-fidelity'' in the sense that the code lacks much of the functionality of the original GEMS solver (e.g. some robustness controls, a wide variety of flow physics and reaction models, parallel processing), the problems it is designed to model are still extremely challenging for modern ROM methods. We hope that open access and contribution to this code can help tackle some of the many problems in the field of ROMs for reacting flows.
 
-This document specifically serves as a reference for the theory behind the unsteady reacting flow solver and ROM methods. A separate user guide is supplied in the documentation directory of the code repository. If you notice any errors in either document (or the code), please create a new issue on the GitHub repository. If you would like to contribute to the code, please fork the repository and submit a pull request when you're finished.
+This document specifically serves as a reference for the theory behind the unsteady reacting flow solver and ROM methods. A separate user guide is supplied at \texttt{https://perform.readthedocs.io}. If you notice any errors in the documentation (or the code), please create a new issue on the GitHub repository. If you would like to contribute to the code, please fork the repository and submit a pull request when you're finished.
 
-The mathematical notation in this paper is as follows: scalars and scalar-valued functions are denoted by italicized, lowercase letters/symbols (e.g., $\dummyVar$), vectors and vector-valued functions are denoted by boldface, lowercase letters/symbols (e.g., \dummyVec), matrices and matrix-valued functions are denoted by boldface, uppercase letters/symbols (e.g., \dummyMat), and vector spaces are denoted by calligraphic, uppercase letters/symbols (e.g., \dummySpace).
+The mathematical notation in this document is as follows: scalars and scalar-valued functions are denoted by italicized, lowercase letters/symbols (e.g., $\dummyVar$), vectors and vector-valued functions are denoted by boldface, lowercase letters/symbols (e.g., \dummyVec), matrices and matrix-valued functions are denoted by boldface, uppercase letters/symbols (e.g., \dummyMat), and vector spaces are denoted by calligraphic, uppercase letters/symbols (e.g., \dummySpace).
 
