+++ .github/workflows/perform_intesive.yml
@@ -7,8 +7,6 @@ on:
   pull_request:
     branches:
       - main
-  schedule:
-    - cron: "0 5 * * *"
 
 
 jobs:
