### 2022-08-31 #3428 @cwentland0
Hi @kyleniemeyer , do you think you'll be able to get to this this week? I only ask because a collaborator is asking to reference the paper if it's finished this week. Otherwise I'll just ask them to cite the Zenodo DOI, so no worries either way!
