### 2022-10-20 251608e
update Elnaz's paper ref, add Wayne's paper ref
