+++ paper/paper.bib
@@ -132,9 +132,11 @@
 
 @article{Rezaian2021,
 	author = {Elnaz Rezaian and Cheng Huang and Karthik Duraisamy},
-	title = {Non-intrusive Balancing Transformation of Highly Stiff Systems with Lightly-damped Impulse Response},
-	year = {2021},
-	journal = {arXiv:2109.10408}
+	title = {Non-intrusive balancing transformation of highly stiff systems with lightly damped impulse response},
+	year = {2022},
+	journal = {Philosophical Transactions of the Royal Society A},
+	volume = {380},
+	doi = {10.1098/rsta.2021.0202},
 }
 
 @article{Everson1995,
@@ -155,4 +157,11 @@
 	volume = {404},
 	doi = {10.1016/j.jcp.2019.108973},
 	year = {2020}
+}
+
+@article{Uy2022
+	author = {Wayne Isaac Tan Uy, Christopher R. Wentland, Cheng Huang, Benjamin Peherstorfer},
+	title = {Reduced models with nonlinear approximations of latent dynamics for model premixed flame problems},
+	year = {2022},
+	journal = {arXiv:2209.06957},
 }
