+++ paper/paper.bib
@@ -159,7 +159,7 @@
 	year = {2020}
 }
 
-@article{Uy2022
+@article{Uy2022,
 	author = {Wayne Isaac Tan Uy, Christopher R. Wentland, Cheng Huang, Benjamin Peherstorfer},
 	title = {Reduced models with nonlinear approximations of latent dynamics for model premixed flame problems},
 	year = {2022},
