### 2022-10-20 b764aba
fix incorrect names listing in Wayne's paper ref
