+++ paper/paper.bib
@@ -160,7 +160,7 @@
 }
 
 @article{Uy2022,
-	author = {Wayne Isaac Tan Uy, Christopher R. Wentland, Cheng Huang, Benjamin Peherstorfer},
+	author = {Wayne Isaac Tan Uy and Christopher R. Wentland and Cheng Huang and Benjamin Peherstorfer},
 	title = {Reduced models with nonlinear approximations of latent dynamics for model premixed flame problems},
 	year = {2022},
 	journal = {arXiv:2209.06957},
