### 2022-10-21 #3428 @kyleniemeyer
Hi @cwentland0, sorry for my (extreme) delay here. I was able to install and run the package, including the tests and examples.

A few comments:
- Your testing approach is not very standard—have you considered setting things up to work with a standard tool like pytest? This would allow automated running of tests. (You can also specify the additional optional/testing requirement of TensorFlow in the `setup.py` file.)
- Related, but some indication that the regression tests are going to take a while would be nice...
- Regarding the command-line interface, have you considered allowing users to specify the input file, rather than requiring them to have a rigid directory/file-naming structure?
- The Python community has moved away from using `setup.py` with `setuptools` to `pyproject.toml` as the standard. I realize you submitted this a while ago, but moving forward you should probably switch to that system.
