### 2022-10-21 #3428 @cwentland0
Hey @kyleniemeyer no worries, thank you for reviewing! Addressing your comments:

- I remember looking into `nose`, `pytest`, and `unittest`, and ended up opting for `unittest` since it seemed the simplest. I can look into using `nose` if that would be considered more standard? From what I can tell that would require less rewriting of the original testing code.
- I apologize for the long regression tests, I will definitely add in a warning when it asks for confirmation to run them!
- Command line input file specification definitely crossed my mind, I can add that in very simply while leaving the current setup as the default if the user doesn't supply an input file path.
- Looks like adding `pyproject.toml` would be fairly simple from a cursory overview. From what I can tell, though, `pyproject.toml` does not permit an editable install like `pip install -e`. Since the idea for PERFORM is that folks can easily expand the code, it would still need an editable install, but from what I'm reading it seems like they can coexist.

I can get the last three things done today, if you would like the tests to be rewritten with `pytest` that will take a bit longer.
