### 2022-10-21 f20ba41
command line option to specify solver and rom params input files
