### 2022-10-22 #3428 @kyleniemeyer
@cwentland0

> I remember looking into nose, pytest, and unittest, and ended up opting for unittest since it seemed the simplest. I can look into using nose if that would be considered more standard? From what I can tell that would require less rewriting of the original testing code.

I think that `pytest` is probably the widest used right now, and is what I would consider the modern "standard". From my understanding folks have moved away from `nose`, actually. To be compatible with `pytest`, you just have to have your test files and the classes/functions inside that start with the word "test", essentially, which shouldn't require *too* much work.
But, I

> I apologize for the long regression tests, I will definitely add in a warning when it asks for confirmation to run them! Command line input file specification definitely crossed my mind, I can add that in very simply while leaving the current setup as the default if the user doesn't supply an input file path.

> Looks like adding pyproject.toml would be fairly simple from a cursory overview. From what I can tell, though, pyproject.toml does not permit an editable install like pip install -e. Since the idea for PERFORM is that folks can easily expand the code, it would still need an editable install, but from what I'm reading it seems like they can coexist.

You actually can still use a `pyproject.toml` approach with an editable install: https://www.python.org/dev/peps/pep-0660/
