### 2022-10-30 5f3c0c6
integrate black and coverage config into pyproject.toml, convert setup.py to setup.cfg
