+++ pyproject.toml
@@ -1,3 +1,24 @@
 [build-system]
-requires = ["setuptools"]
+requires = ["setuptools >= 64.0.0", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[tool.black]
+line-length = 120
+target-version = ['py38']
+check = true
+diff = true
+
+[tool.coverage.report]
+exclude_lines = [
+    "pragma: no cover",
+    "pass",
+	"print",
+	"sleep",
+	"return",
+    "assert",
+    "raise ValueError",
+    "def main()",
+]
+omit = [
+	"*/perform/driver.py",
+]
