+++ setup.cfg
@@ -1,0 +1,24 @@
+[metadata]
+name = perform
+version = 0.1
+author = Christopher R. Wentland
+author_email = chriswen@umich.edu
+url = https://github.com/cwentland0/perform
+description = One-dimension reacting flow for ROM prototyping
+long_description = file: README.md
+license = MIT
+license_files = LICENSE
+
+[options]
+python_requires = >= 3.6
+install_requires =
+	numpy >= 1.16.6
+	scipy >= 1.1.0
+	matplotlib >= 2.1.0
+	packaging
+	gdown
+
+[options.entry_points]
+console_scripts =
+	perform = perform.driver:main
+
