### 2022-10-30 0307acd
move everything from setup.cfg to pyproject.toml
