### 2022-10-30 #3428 @cwentland0
I've merged the changes for the latter three points into the main branch (including replacing `setup.py` with `pyproject.toml`). For now I'm going to keep using `unittest` but will consider using `pytest` moving forward. Please let me know if there's anything else you'd like me to address!
