### 2022-11-08 #3428 @cwentland0
Just updated the `pyproject.toml` and made a tagged release. Actually updated to 0.2.1, as I labelled the last release as 0.2, I just wasn't careful to update it in the setup.py then!
