### 2022-11-08 #3428 @kyleniemeyer
@cwentland0 great! at this point please archive the repository (e.g., on Zenodo) and provide the DOI here.

FYI, moving forward, you might consider encoding the version number in a file (e.g., `_version.py`) and then have `pyproject.toml` read that in dynamically.
