### 2022-11-08 #3428 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
