### 2022-11-08 #3428 @kyleniemeyer
@cwentland0 just a few last changes for the paper: could you add the arXiv DOIs for the two arXiv citations? 10.48550/arXiv.2003.07798 and 10.48550/arXiv.2209.06957

No need to update the Zenodo archive
