### 2022-11-08 59cda30
add DOIs for Pression and Wayne's paper
