+++ paper/paper.bib
@@ -1,8 +1,9 @@
 @article{Rizzi2020,
 	title = {{Pressio: Enabling projection-based model reduction for large-scale nonlinear dynamical systems}},
 	author = {Rizzi, Francesco and Blonigan, Patrick J and Carlberg, Kevin T},
-	journal = {arXiv:2003.07798},
-	year = {2020}
+	journal = {arXiv},
+	year = {2020},
+	doi = {10.48550/arXiv.2003.07798},
 }
 
 @article{Stabile2017CAIM,
@@ -163,5 +164,6 @@
 	author = {Wayne Isaac Tan Uy and Christopher R. Wentland and Cheng Huang and Benjamin Peherstorfer},
 	title = {Reduced models with nonlinear approximations of latent dynamics for model premixed flame problems},
 	year = {2022},
-	journal = {arXiv:2209.06957},
+	journal = {arXiv},
+	doi = {10.48550/arXiv.2209.06957}
 }
