### 2022-11-08 #3428 @kyleniemeyer
Congratulations @cwentland0 on your article's publication in JOSS! And thanks for your patience during this process.

Many thanks to @Himscipy for reviewing.
